Class for representing HDF5 attributes

This class represents an HDF5 attribute. Usually it is easier to read and write attributes for groups, datasets and committed datatypes using the functions documented in h5attributes.

Otherwise, the functionality for attributes is very similar to that of datasets (H5D), however with the notable exception that attributes always have to be read and written as a whole.

Value

Object of class H5A.

Author

Holger Hoefling

Super class

H5RefClass -> H5A

Methods

Public methods

H5A$flush()

H5A$get_filename()

H5A$get_info()

H5A$attr_name()

H5A$get_space()

H5A$get_type()

H5A$get_storage_size()

H5A$read_low_level()

H5A$read()

H5A$write_low_level()

H5A$write()

H5A$print()

Inherited methods

H5RefClass$close()

H5RefClass$dec_ref()

H5RefClass$get_file_id()

H5RefClass$get_obj_type()

H5RefClass$get_ref()

H5RefClass$inc_ref()

H5RefClass$initialize()

H5RefClass$methods()

H5A$flush()

This function implements the HDF5-API function H5Fflush. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html for details.

Usage

H5A$flush(scope = h5const$H5F_SCOPE_GLOBAL)

Arguments

scope

The scope of the flush; one of the H5F_SCOPE_* values in h5const.

H5A$get_filename()

This function implements the HDF5-API function H5Fget_name. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_f.html for details.

Usage

H5A$get_filename()

H5A$get_info()

This function implements the HDF5-API function H5Aget_info. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5A$get_info()

H5A$attr_name()

This function implements the HDF5-API function H5Aget_name. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5A$attr_name()

H5A$get_space()

This function implements the HDF5-API function H5Aget_space. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5A$get_space()

H5A$get_type()

This function implements the HDF5-API function H5Aget_type. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5A$get_type(native = TRUE)

Arguments

native

Logical; should the native version of the datatype be returned.

H5A$get_storage_size()

This function implements the HDF5-API function H5Aget_storage_size. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5A$get_storage_size()

H5A$read_low_level()

Only for advanced users. See documentation for read instead. This function implements the HDF5-API function H5Aread. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5A$read_low_level(buffer, mem_type, duplicate_buffer = FALSE)

Arguments

buffer

The raw buffer to read into or write from.

mem_type

The memory datatype to use for the transfer; an object of class H5T.

duplicate_buffer

Logical; should the buffer be copied before it is returned.

H5A$read()

Reads the data of the attribute and returns it as an R-object

Usage

H5A$read(flags = getOption("hdf5r.h5tor_default"), drop = TRUE)

Arguments

flags

Conversion rules for integer values. See also h5const

drop

Logical. Should dimensions of length 1 be dropped (R-default for arrays)

H5A$write_low_level()

Only for advanced users. See documentation for write instead. This function implements the HDF5-API function H5Awrite. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5A$write_low_level(buffer, mem_type)

Arguments

buffer

The raw buffer to read into or write from.

mem_type

The memory datatype to use for the transfer; an object of class H5T.

H5A$write()

Writes the data of robj to the attribute

Usage

H5A$write(robj, mem_type = NULL, flush = getOption("hdf5r.flush_on_write"))

Arguments

robj

The object to write into the attribute

mem_type

The memory data type to use when transferring from HDF5 to intermediate storage. This is an advanced development feature and may be removed in the future.

flush

Logical; should the file be flushed after the write.

H5A$print()

Prints information for the dataset

Usage

H5A$print(...)

Arguments

...

ignored

Examples

fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
h5attr(file, "attr_numeric") <- rnorm(10)
a <- file$attr_open("attr_numeric")
a$get_info()
#>   corder_valid corder cset data_size
#> 1            0      0    0        80
a$attr_name()
#> [1] "attr_numeric"
a$get_space()
#> Class: H5S
#> Type: Simple
#> Dims: 10
#> Maxdims: 10
a$get_type()
#> Class: H5T_FLOAT
#> Datatype: H5T_IEEE_F64LE
a$get_storage_size()
#> [1] 80
a$read()
#>  [1] -1.400043517  0.255317055 -2.437263611 -0.005571287  0.621552721
#>  [6]  1.148411606 -1.821817661 -0.247325302 -0.244199607 -0.282705449
a$write(10:1)
a$print()
#> Class: H5A
#> Attribute: attr_numeric
#> Datatype: H5T_IEEE_F64LE
#> Space: Type=Simple     Dims=10     Maxdims=10
a$close()
file$close_all()