Class for representing HDF5 groups

HDF5-Groups are essentially equivalent to directories in a file system. Inside the groups, other groups or datasets can be created. For the most parts, groups behave like files, so please also look at the documentation of H5File.

Value

Object of class H5Group.

Methods

print(..., max.attributes = 10, max.listing = 10)

Prints information for the group Parameters

max.attributes

Maximum number of attribute names to print

max.listing

Maximum number of ls-items to print

...

ignored

open(name, link_access_pl = h5const$H5P_DEFAULT, dataset_access_pl = h5const$H5P_DEFAULT, type_access_pl = h5const$H5P_DEFAULT)

Opens groups, datasets or types using the appropriate HDF5-API functions. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5D.html#Dataset-Open for datasets, https://www.hdfgroup.org/HDF5/doc/RM/RM_H5O.html#Object-Open for types and https://www.hdfgroup.org/HDF5/doc/RM/RM_H5O.html#Object-Open for general objects.

open_by_idx(n, group_name = ".", index_type = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Oopen_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5O.html#Object-OpenByIdx for details.

ls(recursive = FALSE, detailed = FALSE, index_type = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT, dataset_access_pl = h5const$H5P_DEFAULT, type_access_pl = h5const$H5P_DEFAULT)

Returns the contents of a file or group as a data.frame.

exists(name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lexists. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-Exists for details.

path_valid(path, check_object_valid = TRUE)

This function implements the HDF5-API function H5LTpath_valid. Please see the documentation at https://support.hdfgroup.org/HDF5/doc/HL/RM_H5LT.html#H5LTpath_valid for details.

link(obj, new_link_name, link_create_pl = h5const$H5P_DEFAULT, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Olink. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5O.html#Object-Link for details.

obj_copy_to(dst_loc, dst_name, src_name, object_copy_pl = h5const$H5P_DEFAULT, link_create_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Ocopy. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5O.html#Object-Copy for details.

obj_copy_from(src_loc, src_name, dst_name, object_copy_pl = h5const$H5P_DEFAULT, link_create_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Ocopy. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5O.html#Object-Copy for details.

obj_info_by_idx(n, group_name = ".", index_field = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, remove_internal_use_only = TRUE)

This function implements the HDF5-API function H5Oget_info_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5O.html#Object-GetInfoByIdx for details.

obj_info_by_name(object_name, remove_internal_use_only = TRUE)

This function implements the HDF5-API function H5Oget_info_by_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5O.html#Object-GetInfoByName for details.

group_info()

This function implements the HDF5-API function H5Gget_info. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5G.html#Group-GetInfo for details.

group_info_by_name(name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Gget_info_by_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5G.html#Group-GetInfoByName for details.

group_info_by_idx(n, group_name = ".", index_field = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Gget_info_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5G.html#Group-GetInfoByIdx for details.

create_group(name, link_create_pl = h5const$H5P_DEFAULT, group_create_pl = h5const$H5P_DEFAULT, group_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Gcreate2 and H5Gcreate_anon (if name is NULL). Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5G.html#Group-Create2 for regular groups and https://www.hdfgroup.org/HDF5/doc/RM/RM_H5G.html#Group-CreateAnon for anonymous groups for details.

create_dataset(name, robj = NULL, dtype = NULL, space = NULL, dims = NULL, chunk_dims = "auto", gzip_level = 4, link_create_pl = h5const$H5P_DEFAULT, dataset_create_pl = h5const$H5P_DEFAULT, dataset_access_pl = h5const$H5P_DEFAULT)

This function is the main interface to create a new dataset. Its parameters allow for customization of the default behavior, i.e. in order to get a specific datatype, a certain chunk size or dataset dimensionality. Also note that this function implements the HDF5-API function H5Dcreate2 and H5Dcreate_anon (if name is NULL). Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5D.html#Dataset-Create2 for regular groups and https://www.hdfgroup.org/HDF5/doc/RM/RM_H5D.html#Dataset-CreateAnon for anonymous groups for details. Parameters

name

The name of the new dataset. If missing, an anonymous dataset is created

robj

An R-object to take as a template for creating the dataset. Either robj or both dtype and space have to be provided

dtype

The datatype to use for the creation of the object. Can be null if robj is given.

space

The space to use for the object creation. Can be null if robj is given. Otherwise an object of type H5S which specifies the dimensions of the dataset.

dims

Dimension of the new dataset; used if space is NULL. overwrite the dimension guessed from robj if robj is given.

chunk_dims

Size of the chunk. Has to have the same length as the dataset dimension. If "auto" then the size of each chunk is estimated so that each chunk is roughly as large in bytes as the value in the hdf5r.chunk_size option. See also guess_chunks for a more detailed explanation. If set to NULL, then no chunking is used, unless explicitly set in dataset_create_pl.

gzip_level

Only if chunk_dims is not null. If given, then the dataset_create_pl is set to enable zipping at the level given here. If set to NULL, then gzip is not set (but could be set otherwise in dataset_create_pl

link_create_pl

Link creation property list. See H5P_LINK_CREATE

dataset_create_pl

Dataset creation property list. See H5P_DATASET_CREATE

dataset_access_pl

Dataset access property list. See H5P_DATASET_ACCESS

commit(name, dtype, link_create_pl = h5const$H5P_DEFAULT, type_create_pl = h5const$H5P_DEFAULT, type_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Tcommit2. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5T.html#Datatype-Commit2 for details.

link_create_hard(obj_loc, obj_name, link_name, link_create_pl = h5const$H5P_DEFAULT, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lcreate_hard. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-CreateHard for details.

link_create_soft(target_path, link_name, link_create_pl = h5const$H5P_DEFAULT, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lcreate_soft. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-CreateSoft for details.

link_create_external(target_filename, target_obj_name, link_name, link_create_pl = h5const$H5P_DEFAULT, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lcreate_external. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-CreateExternal for details.

link_exists(name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lexists. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-Exists for details.

link_move_from(src_loc, src_name, dst_name, link_create_pl = h5const$H5P_DEFAULT, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lmove. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-Move for details.

link_move_to(dst_loc, dst_name, src_name, link_create_pl = h5const$H5P_DEFAULT, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lmove. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-Move for details.

link_copy_from(src_loc, src_name, dst_name, link_create_pl = h5const$H5P_DEFAULT, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lcopy. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-Copy for details.

link_copy_to(dst_loc, dst_name, src_name, link_create_pl = h5const$H5P_DEFAULT, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lcopy. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-Copy for details.

link_delete(name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Ldelete. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-Delete for details.

link_delete_by_idx(n, group_name = ".", index_field = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Ldelete_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-DeleteByIdx for details.

link_info(name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lget_info. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-GetInfo for details.

link_info_by_idx(n, group_name = ".", index_field = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lget_info_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-GetInfoByIdx for details.

link_value(name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lget_val. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-GetVal for details.

link_value_by_idx(n, group_name = ".", index_field = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lget_val_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-GetValByIdx for details.

link_name_by_idx(n, group_name, idx_type = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Lget_name_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-GetNameByIdx for details.

mount(name, child)

This function implements the HDF5-API function H5Fmount. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5F.html#File-Mount for details.

unmount(name)

This function implements the HDF5-API function H5Funmount. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5F.html#File-Unmount for details.

create_reference(name = ".", space = NULL)

This function implements the HDF5-API function H5Rcreate. If space=NULL then a H5R_OBJECT reference is created, otherwise a H5R_DATASET_REGION reference Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5R.html#Reference-Create for details.

obj_info(remove_internal_use_only = TRUE)

This function implements the HDF5-API function H5Oget_info. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5O.html#Object-GetInfo for details.

get_obj_name()

This function implements the HDF5-API function H5Iget_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5I.html#Identify-GetName for details.

create_attr(attr_name, robj = NULL, dtype = NULL, space = NULL)

This function implements the HDF5-API function H5Acreate2. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-Create2 for details.

attr_open(attr_name)

This function implements the HDF5-API function H5Aopen. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-Open for details.

create_attr_by_name(attr_name, obj_name, robj = NULL, dtype = NULL, space = NULL, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Acreate_by_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-CreateByName for details.

attr_open_by_name(attr_name, obj_name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Aopen_by_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-OpenByName for details.

attr_open_by_idx(n, obj_name, idx_type = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Aopen_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-OpenByIdx for details.

attr_exists_by_name(attr_name, obj_name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Aexists_by_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-ExistsByName for details.

attr_exists(attr_name)

This function implements the HDF5-API function H5Aexists. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-Exists for details.

attr_rename_by_name(old_attr_name, new_attr_name, obj_name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Arename_by_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-RenameByName for details.

attr_rename(old_attr_name, new_attr_name)

This function implements the HDF5-API function H5Arename. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-Rename for details.

attr_delete(attr_name)

This function implements the HDF5-API function H5Adelete. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-Delete for details.

attr_delete_by_name(attr_name, obj_name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Adelete_by_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-DeleteByName for details.

attr_delete_by_idx(n, obj_name, idx_type = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Adelete_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-DeleteByIdx for details.

attr_info_by_name(attr_name, obj_name, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Aget_info_by_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-GetInfoByName for details.

attr_info_by_idx(n, obj_name, idx_type = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Aget_info_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-GetInfoByIdx for details.

attr_name_by_idx(n, obj_name, idx_type = h5const$H5_INDEX_NAME, order = h5const$H5_ITER_NATIVE, link_access_pl = h5const$H5P_DEFAULT)

This function implements the HDF5-API function H5Aget_name_by_idx. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-GetNameByIdx for details.

attr_get_number()

This function implements the HDF5-API function H5Aget_num_attrs. Please see the documentation at https://support.hdfgroup.org/HDF5/doc/RM/RM_H5A.html#Annot-NumAttrs for details.

flush(scope = h5const$H5F_SCOPE_GLOBAL)

This function implements the HDF5-API function H5Fflush. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5F.html#File-Flush for details.

get_filename()

This function implements the HDF5-API function H5Fget_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5F.html#File-GetName for details.

names(link_access_pl = h5const$H5P_DEFAULT)

Returns the names of the items in the group or at the root of the file Parameters

link_access_pl

The link-access property list. See https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html for more detail.

Examples

fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
group <- file$create_group("testgroup")
group$print()

#> Class: H5Group
#> Filename: /tmp/RtmpCQ9cIm/file12507e80c1cc.h5
#> Group: /testgroup

group$close()
file$close_all()