Retrieve object from a group of file

Retrieve object from a group of file

# S3 method for H5Group
[[(x, name, ..., link_access_pl = h5const$H5P_DEFAULT,
  dataset_access_pl = h5const$H5P_DEFAULT,
  type_access_pl = h5const$H5P_DEFAULT)

# S3 method for H5File
[[(x, name, ..., link_access_pl = h5const$H5P_DEFAULT,
  dataset_access_pl = h5const$H5P_DEFAULT,
  type_access_pl = h5const$H5P_DEFAULT)

# S3 method for H5Group
[[(x, name, ...) <- value

# S3 method for H5File
[[(x, name, ...) <- value

Arguments

x
An object of class H5File or H5Group

name
Name of the object to retrieve. Has to be a character vector of length one. No integer values allowed.

...
Currently ignored

link_access_pl
An object of class H5P_LINK_ACCESS-class.

dataset_access_pl
An object of class H5P_DATASET_ACCESS-class.

type_access_pl
Currently always h5const$H5P_DEFAULT

value
What to assign. Has to be one of H5Group-class, H5D-class or H5T-class

Value

A H5Group-class, H5D-class or H5T-class, depending on the object saved in the group under the requested name.

Details

Works similar to retrieving objects in a list. x[["my_name"]] retrieves object my_name from the HDF5-File or HDF5-Group x.

One can also assign objects under a not yet existing name. For either a H5Group-class or H5D-class, a hard link is created. If it is a datatype, H5T-class, this is committed under the chosen name name.