Class for HDF5 property list for file creation

It inherits all functions of the H5P.

Value

Object of class H5P_FILE_ACCESS.

See also

Author

Holger Hoefling

Super classes

H5RefClass -> H5P -> H5P_FILE_ACCESS

Methods

Public methods

H5P_FILE_ACCESS$new()

H5P_FILE_ACCESS$set_cache()

H5P_FILE_ACCESS$get_cache()

Inherited methods

H5RefClass$close()

H5RefClass$dec_ref()

H5RefClass$get_file_id()

H5RefClass$get_obj_type()

H5RefClass$get_ref()

H5RefClass$inc_ref()

H5RefClass$methods()

H5RefClass$print()

H5P$copy()

H5P$equal()

H5P$get_class()

H5P$get_class_name()

H5P_FILE_ACCESS$new()

Create a new class of type H5P_FILE_ACCESS

Usage

H5P_FILE_ACCESS$new(id = NULL)

Arguments

id

Internal use only

H5P_FILE_ACCESS$set_cache()

This function implements the HDF5-API function H5Pset_cache. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html for details.

Usage

H5P_FILE_ACCESS$set_cache(
  rdcc_nslots = 521,
  rdcc_nbytes = 2^20,
  rdcc_w0 = 0.75
)

Arguments

rdcc_nslots

The number of slots in the raw data chunk cache. Use -1 for the library default.

rdcc_nbytes

The size of the raw data chunk cache in bytes. Use -1 for the library default.

rdcc_w0

The chunk preemption policy, between 0 and 1. Use -1 for the library default.

H5P_FILE_ACCESS$get_cache()

This function implements the HDF5-API function H5Pget_cache. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_p.html for details.

Usage

H5P_FILE_ACCESS$get_cache()