Class for HDF5 Reference datatypes.

H5R is only the common base class and is never used. User should not create objects of this class by themselves and instead use the create_reference methods of H5D, H5Group or H5File classes. Sub-classes are H5R_OBJECT and H5R_DATASET_REGION

Value

Object of class H5R.

Author

Holger Hoefling

Super class

H5RefClass -> H5R

Active bindings

length

Get the length of the object

ref

Get or assign the internal raw-vector representation of the data. Usually, user's shouldn't have to use this.

dim

Get or assign the dimensionality of the object

dimnames

Get or assign the dimnames of the object

names

Get or assign the names of the object

rank

Get the rank of the object

Methods

Public methods

H5R$subset_read()

H5R$subset2_read()

H5R$subset_assign()

H5R$subset2_assign()

H5R$t()

H5R$clone()

Inherited methods

H5RefClass$close()

H5RefClass$dec_ref()

H5RefClass$get_file_id()

H5RefClass$get_obj_type()

H5RefClass$get_ref()

H5RefClass$inc_ref()

H5RefClass$initialize()

H5RefClass$methods()

H5RefClass$print()

H5R$subset_read()

Method that returns a subset of the data in the H5R-object

Usage

H5R$subset_read(dim_index, drop = TRUE)

Arguments

dim_index

A list of dimension indices as usually pasted into [

drop

Logical. Should dimensions of size 1 be dropped.

H5R$subset2_read()

Method to read a single item

Usage

H5R$subset2_read(i, exact = TRUE)

Arguments

i

The single item to read

exact

Is the item name exact or should partial matching be allowed?

H5R$subset_assign()

Assign values into a subset of the H5R-vector

Usage

H5R$subset_assign(dim_index, value)

Arguments

dim_index

A list of dimension indices as usually passed into [

value

The value to assign

H5R$subset2_assign()

Assign a value to a single value in the array

Usage

H5R$subset2_assign(i, exact = TRUE, value)

Arguments

i

the index where to assign the value

exact

Logical; does the subscript have to match exactly.

value

The value to assign

H5R$t()

Transpose the object if it is a matrix (i.e. has rank 2

Usage

H5R$t()

H5R$clone()

The objects of this class are cloneable with this method.

Usage

H5R$clone()

Examples

fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
file[["testset"]] <- matrix(rnorm(9), nrow = 3)
dset <- file[["testset"]]
r <- file$create_reference("testset")
file$close_all()