Class for HDF5 dataset-region references.

H5R_DATASET_REGION is the reference class for dataset regions. Users should not create this class by themselves, but use the appropriate and instead use the create_reference methods of H5D, H5Group or H5File classes.

Value

Object of class H5R_DATASET_REGION.

Methods

new(num = 0, id = NULL)

Create a new reference for dataset regions; Usually, users shouldn't have to call this, but use the create_reference method of a dataset.

dereference(object_access_pl = h5const$H5P_DEFAULT, obj = NULL, get_value = FALSE)

Dereference an H5R reference for a dataset region. The file the reference is pointing to is assigned automatically. It returns a list where each item is a list with components dataset, being an H5D object and space being a H5S object. When setting get_value=TRUE, then instead of these objects The data itself is returned This function implements the HDF5-API function H5Rdereference. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5R.html#Reference-Dereference for details. Parameters

obj

Overriding the default file the reference is referring to

object_access_pl

The object-access property list. Currently always the default