Base class that tracks the ids and allows for closing an id

This class is not intended for creating objects, but as a base class for all other H5-derived classes to provide common functionality for id tracking

Value

Object of reference class H5RefClass.

Author

Holger Hoefling

Active bindings

id

Returns the id of the object

id_env

The environment that holds the id, used internally for reference counting.

is_valid

This function implements the HDF5-API function H5Iis_valid. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html for details. Additionally, the R-object representing the HDF5-id can be invalidated as well. In this case, the class id is set to NA and is_valid returns FALSE.

message

Legacy function; currently not used; may be removed

Methods

Public methods

H5RefClass$get_file_id()

H5RefClass$get_obj_type()

H5RefClass$get_ref()

H5RefClass$inc_ref()

H5RefClass$dec_ref()

H5RefClass$new()

H5RefClass$close()

H5RefClass$print()

H5RefClass$methods()

H5RefClass$get_file_id()

This function implements the HDF5-API function H5Iget_file_id. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html for details.

Usage

H5RefClass$get_file_id()

H5RefClass$get_obj_type()

This function implements the HDF5-API function H5Iget_type. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html for details.

Usage

H5RefClass$get_obj_type()

H5RefClass$get_ref()

This function implements the HDF5-API function H5Iget_ref. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html for details.

Usage

H5RefClass$get_ref()

H5RefClass$inc_ref()

This function implements the HDF5-API function H5Iinc_ref. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html for details.

Usage

H5RefClass$inc_ref()

H5RefClass$dec_ref()

This function implements the HDF5-API function H5Idec_ref. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html for details.

Usage

H5RefClass$dec_ref()

H5RefClass$new()

Constructor for the basic class for hdf5 objects. Takes an id and stores it appropriately, including the necessary counting of object references that the package implements. This reference counting is included in addition to R's internal method in order to allow for the invalidation of objects in R itself when all open handles in an R-file are being closed.

Usage

H5RefClass$new(id = NULL)

Arguments

id

An HDF5 id; for internal use only.

H5RefClass$close()

Closes an object and calls the appropriate HDF5 function for the type of object

Usage

H5RefClass$close()

H5RefClass$print()

Prints the class of the object and the id

Usage

H5RefClass$print(...)

Arguments

...

Further arguments; currently ignored.

H5RefClass$methods()

Prints available methods on the screen

Usage

H5RefClass$methods()