Base class that tracks the ids and allows for closing an id

This class is not intended for creating objects, but as a base class for all other H5-derived classes to provide common functionality for id tracking

Value

Object of reference class H5RefClass.

Fields

id

Returns the id of the object as an integer

Methods

new(id = NULL)

Constructor for the basic class for hdf5 objects. Takes an id and stores it appropriately, including the necessary counting of object references that the package implements. This reference counting is included in addition to R's internal method in order to allow for the invalidation of objects in R itself when all open handles in an R-file are being closed.

close()

Closes an object and calls the appropriate HDF5 function for the type of object

print(...)

Prints the class of the object and the id

methods()

Prints available methods on the screen

get_file_id()

This function implements the HDF5-API function H5Iget_file_id. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5I.html#Identify-GetFileId for details.

get_obj_type()

This function implements the HDF5-API function H5Iget_type. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5I.html#Identify-GetType for details.

get_ref()

This function implements the HDF5-API function H5Iget_ref. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5I.html#Identify-GetRef for details.

inc_ref()

This function implements the HDF5-API function H5Iinc_ref. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5I.html#Identify-IncRef for details.

dec_ref()

This function implements the HDF5-API function H5Idec_ref. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5I.html#Identify-DecRef for details.

id()

Returns the id of the object

is_valid()

This function implements the HDF5-API function H5Iis_valid. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5I.html#Identify-IsValid for details. Additionally, the R-object representing the HDF5-id can be invalidated as well. In this case, the class id is set to NA and is_valid returns FALSE.

message()

Legacy function; currently not used; may be removed