Class for representing HDF5 spaces

This class represents Spaces in HDF5. These are mostly useful to define the dimensions of a dataset as well as the maximum dimensions to which it can grow. By default, the maximum dimension is equal to the initial dimension. If you want the array to be able to grow arbitrarily large in one dimension, set the maximum dimension for this index to Inf. See the examples below for code how to do this.

Value

Object of class H5S.

Author

Holger Hoefling

Super class

H5RefClass -> H5S

Active bindings

dims

Get the dimensions of the space. Return NULL if the space is not simple (i.e. NULL-space) or a length-0 integer if it is a scalar

maxdims

Get the maximal dimensions of the space.Return NULL if the space is not simple (i.e. NULL-space) or a length-0 integer if it is a scalar

rank

The rank, i.e. the number of dimensions, of the space.

Methods

Public methods

H5S$new()

H5S$copy()

H5S$encode()

H5S$is_simple()

H5S$get_simple_extent_ndims()

H5S$offset_simple()

H5S$get_simple_extent_dims()

H5S$get_simple_extent_npoints()

H5S$get_simple_extent_type()

H5S$extent_copy()

H5S$extent_equal()

H5S$set_extent_simple()

H5S$set_extent_none()

H5S$get_select_type()

H5S$get_select_npoints()

H5S$get_select_hyper_nblocks()

H5S$get_select_hyper_blocklist()

H5S$get_select_elem_npoints()

H5S$get_select_elem_pointlist()

H5S$get_select_bounds()

H5S$select_all()

H5S$select_none()

H5S$select_valid()

H5S$select_elements()

H5S$select_hyperslab()

H5S$subset()

H5S$print()

Inherited methods

H5RefClass$close()

H5RefClass$dec_ref()

H5RefClass$get_file_id()

H5RefClass$get_obj_type()

H5RefClass$get_ref()

H5RefClass$inc_ref()

H5RefClass$methods()

H5S$new()

Create a new HDF5-space. This can be done by either specifying a space with appropriate dimensions or by decoding a character string that represents an encoded space

Usage

H5S$new(
  type = c("simple", "scalar", "null"),
  dims = NULL,
  maxdims = dims,
  decode_buf = NULL,
  id = NULL
)

Arguments

type

Either a simple space, for which dims and maxdims have to be given or a scalar or null space. See the HDF5 user guide on spaces to explain the differences.

dims

The dimension of the space in case it is of type simple

maxdims

The maximal dimensions of the space

decode_buf

The character string that holds the encoded representation of a space

id

An existing HDF5 id; internal use only

H5S$copy()

This function implements the HDF5-API function H5Scopy. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$copy()

H5S$encode()

This function implements the HDF5-API function H5Sencode. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$encode()

H5S$is_simple()

This function implements the HDF5-API function H5Sis_simple. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$is_simple()

H5S$get_simple_extent_ndims()

This function implements the HDF5-API function H5Sget_simple_extent_ndims. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_simple_extent_ndims()

H5S$offset_simple()

This function implements the HDF5-API function H5Soffset_simple. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$offset_simple(offset)

Arguments

offset

The offset to apply; see the corresponding HDF5 function for how it is interpreted.

H5S$get_simple_extent_dims()

This function implements the HDF5-API function H5Sget_simple_extent_dims. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_simple_extent_dims()

H5S$get_simple_extent_npoints()

This function implements the HDF5-API function H5Sget_simple_extent_npoints. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_simple_extent_npoints()

H5S$get_simple_extent_type()

This function implements the HDF5-API function H5Sget_simple_extent_type. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_simple_extent_type()

H5S$extent_copy()

This function implements the HDF5-API function H5Sextent_copy. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$extent_copy(h5s_source)

Arguments

h5s_source

The dataspace to copy the extent from; an object of class H5S.

H5S$extent_equal()

This function implements the HDF5-API function H5Sextent_equal. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$extent_equal(h5s_cmp)

Arguments

h5s_cmp

The dataspace to compare the extent against; an object of class H5S.

H5S$set_extent_simple()

This function implements the HDF5-API function H5Sset_extent_simple. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$set_extent_simple(dims, maxdims)

Arguments

dims

The dimensions of the object.

maxdims

The maximum dimensions of the dataspace; use Inf for unlimited.

H5S$set_extent_none()

This function implements the HDF5-API function H5Sset_extent_none. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$set_extent_none()

H5S$get_select_type()

This function implements the HDF5-API function H5Sget_select_type. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_select_type()

H5S$get_select_npoints()

This function implements the HDF5-API function H5Sget_select_npoints. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_select_npoints()

H5S$get_select_hyper_nblocks()

This function implements the HDF5-API function H5Sget_select_hyper_nblocks. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_select_hyper_nblocks()

H5S$get_select_hyper_blocklist()

This function implements the HDF5-API function H5Sget_select_hyper_blocklist. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_select_hyper_blocklist(
  startblock = 0,
  numblocks = (self$get_select_hyper_nblocks() - startblock)
)

Arguments

startblock

The first block to return; zero-based.

numblocks

The number of blocks to return.

H5S$get_select_elem_npoints()

This function implements the HDF5-API function H5Sget_select_elem_npoints. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_select_elem_npoints()

H5S$get_select_elem_pointlist()

This function implements the HDF5-API function H5Sget_select_elem_pointlist. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_select_elem_pointlist(
  startpoint = 0,
  numpoints = (self$get_select_elem_npoints() - startpoint)
)

Arguments

startpoint

The first point to return; zero-based.

numpoints

The number of points to return.

H5S$get_select_bounds()

This function implements the HDF5-API function H5Sget_select_bounds. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$get_select_bounds()

H5S$select_all()

This function implements the HDF5-API function H5Sselect_all. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$select_all()

H5S$select_none()

This function implements the HDF5-API function H5Sselect_none. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$select_none()

H5S$select_valid()

This function implements the HDF5-API function H5Sselect_valid. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$select_valid()

H5S$select_elements()

This function implements the HDF5-API function H5Sselect_elements. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$select_elements(coord, op = h5const$H5S_SELECT_SET, byrow = TRUE)

Arguments

coord

The coordinates of the elements to select.

op

The selection operator to apply; one of the H5S_SELECT_* values in h5const.

byrow

Logical; are the coordinates given row-wise.

H5S$select_hyperslab()

This function implements the HDF5-API function H5Sselect_hyperslab. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_s.html for details.

Usage

H5S$select_hyperslab(
  start,
  count,
  stride = NULL,
  block = NULL,
  op = h5const$H5S_SELECT_SET
)

Arguments

start

The offset at which the hyperslab starts, one value per dimension.

count

The number of blocks to select, one value per dimension.

stride

The distance between successive blocks, one value per dimension.

block

The size of each block, one value per dimension.

op

The selection operator to apply; one of the H5S_SELECT_* values in h5const.

H5S$subset()

Subsetting the space. This is mainly intended as a helper function for the '[' function, but can also be used on its own.

Usage

H5S$subset(args, op = h5const$H5S_SELECT_SET, envir = parent.frame())

Arguments

args

The indices for each dimension to subset given as a list. This makes this easier to use as a programmatic API. For interactive use we recommend the use of the [ operator.

op

The operator to use. Same as for the other HDF5 space selection functions. One of the elements shown in h5const$H5S_seloper_t

envir

The environment in which to evaluate args

H5S$print()

Prints information for the group

Usage

H5S$print(...)

Arguments

...

ignored

Examples

h5s_fixed <- H5S$new("simple", dims=c(5, 2))
h5s_fixed
#> Class: H5S
#> Type: Simple
#> Dims: 5 x 2
#> Maxdims: 5 x 2

h5s_variable <- H5S$new("simple", dims=c(5,2), maxdims=c(Inf,2))
h5s_variable
#> Class: H5S
#> Type: Simple
#> Dims: 5 x 2
#> Maxdims: Inf x 2
h5s_variable$set_extent_simple(c(10,2), c(Inf, 2))
h5s_variable
#> Class: H5S
#> Type: Simple
#> Dims: 10 x 2
#> Maxdims: Inf x 2

# now select a subset of points
# argument evaluation has a heuristic; here it chooses point selection
h5s_variable[c(1, 3, 8), 1]
#> Class: H5S
#> Type: Simple
#> Dims: 10 x 2
#> Maxdims: Inf x 2
h5s_variable$get_select_type()
#> [1] H5S_SEL_POINTS
#> 6 Levels: H5S_SEL_ERROR H5S_SEL_NONE H5S_SEL_POINTS ... H5S_SEL_N
#> 6 Values: -1 0 1 ... 4
h5s_variable$get_select_elem_pointlist()
#>      [,1] [,2]
#> [1,]    1    1
#> [2,]    3    1
#> [3,]    8    1

# and a hyperslab (chosen by the argument heuristic)
h5s_variable[2:7, 1:2]
#> Class: H5S
#> Type: Simple
#> Dims: 10 x 2
#> Maxdims: Inf x 2
h5s_variable$get_select_type()
#> [1] H5S_SEL_HYPERSLABS
#> 6 Levels: H5S_SEL_ERROR H5S_SEL_NONE H5S_SEL_POINTS ... H5S_SEL_N
#> 6 Values: -1 0 1 ... 4
h5s_variable$get_select_hyper_blocklist()
#>               [,1] [,2]
#> block_1_start    2    1
#> block_1_end      7    2