Class for HDF5 datatypes.

This is the base class for all datatypes, but most have a specialised class. This class represents an HDF5 datatype. It inherits all functions of the H5RefClass. It is also the base class for many other classes well, specifically

Integer

H5T_INTEGER

Bitfield

H5T_BITFIELD (currently identical to the integer class)

Float

H5T_FLOAT

Enum

H5T_ENUM

Compound

H5T_COMPOUND

String

H5T_STRING

Complex

H5T_COMPLEX

Array

H5T_ARRAY

Variable Length

H5T_VLEN

Value

Object of class H5T.

Author

Holger Hoefling

Super class

H5RefClass -> H5T

Methods

Public methods

H5T$get_obj_name()

H5T$create_attr()

H5T$attr_open()

H5T$create_attr_by_name()

H5T$attr_open_by_name()

H5T$attr_open_by_idx()

H5T$attr_exists_by_name()

H5T$attr_exists()

H5T$attr_rename_by_name()

H5T$attr_rename()

H5T$attr_delete()

H5T$attr_delete_by_name()

H5T$attr_delete_by_idx()

H5T$attr_info_by_name()

H5T$attr_info_by_idx()

H5T$attr_name_by_idx()

H5T$attr_get_number()

H5T$create_reference()

H5T$new()

H5T$get_class()

H5T$get_size()

H5T$set_size()

H5T$set_precision()

H5T$get_precision()

H5T$set_order()

H5T$get_order()

H5T$set_offset()

H5T$get_offset()

H5T$set_pad()

H5T$get_pad()

H5T$copy()

H5T$is_committed()

H5T$equal()

H5T$is_vlen()

H5T$detect_class()

H5T$get_native_type()

H5T$get_create_plist()

H5T$to_text()

H5T$print()

H5T$obj_info()

Inherited methods

H5RefClass$close()

H5RefClass$dec_ref()

H5RefClass$get_file_id()

H5RefClass$get_obj_type()

H5RefClass$get_ref()

H5RefClass$inc_ref()

H5RefClass$methods()

H5T$get_obj_name()

This function implements the HDF5-API function H5Iget_name. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_i.html for details.

Usage

H5T$get_obj_name()

H5T$create_attr()

This function implements the HDF5-API function H5Acreate2. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$create_attr(attr_name, robj = NULL, dtype = NULL, space = NULL)

Arguments

attr_name

The name of the attribute.

robj

An R object to use as a template for the datatype and dataspace.

dtype

The datatype to use; an object of class H5T.

space

The dataspace to use; an object of class H5S.

H5T$attr_open()

This function implements the HDF5-API function H5Aopen. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_open(attr_name)

Arguments

attr_name

The name of the attribute.

H5T$create_attr_by_name()

This function implements the HDF5-API function H5Acreate_by_name. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$create_attr_by_name(
  attr_name,
  obj_name,
  robj = NULL,
  dtype = NULL,
  space = NULL,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

attr_name

The name of the attribute.

obj_name

The name of the object, relative to the object the method is called on.

robj

An R object to use as a template for the datatype and dataspace.

dtype

The datatype to use; an object of class H5T.

space

The dataspace to use; an object of class H5S.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_open_by_name()

This function implements the HDF5-API function H5Aopen_by_name. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_open_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

attr_name

The name of the attribute.

obj_name

The name of the object, relative to the object the method is called on.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_open_by_idx()

This function implements the HDF5-API function H5Aopen_by_idx. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_open_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

n

The index of the item to access; zero-based.

obj_name

The name of the object, relative to the object the method is called on.

idx_type

The index to iterate over; one of the H5_INDEX_* values in h5const.

order

The order in which to iterate; one of the H5_ITER_* values in h5const.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_exists_by_name()

This function implements the HDF5-API function H5Aexists_by_name. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_exists_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

attr_name

The name of the attribute.

obj_name

The name of the object, relative to the object the method is called on.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_exists()

This function implements the HDF5-API function H5Aexists. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_exists(attr_name)

Arguments

attr_name

The name of the attribute.

H5T$attr_rename_by_name()

This function implements the HDF5-API function H5Arename_by_name. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_rename_by_name(
  old_attr_name,
  new_attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

old_attr_name

The current name of the attribute.

new_attr_name

The new name of the attribute.

obj_name

The name of the object, relative to the object the method is called on.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_rename()

This function implements the HDF5-API function H5Arename. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_rename(old_attr_name, new_attr_name)

Arguments

old_attr_name

The current name of the attribute.

new_attr_name

The new name of the attribute.

H5T$attr_delete()

This function implements the HDF5-API function H5Adelete. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_delete(attr_name)

Arguments

attr_name

The name of the attribute.

H5T$attr_delete_by_name()

This function implements the HDF5-API function H5Adelete_by_name. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_delete_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

attr_name

The name of the attribute.

obj_name

The name of the object, relative to the object the method is called on.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_delete_by_idx()

This function implements the HDF5-API function H5Adelete_by_idx. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_delete_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

n

The index of the item to access; zero-based.

obj_name

The name of the object, relative to the object the method is called on.

idx_type

The index to iterate over; one of the H5_INDEX_* values in h5const.

order

The order in which to iterate; one of the H5_ITER_* values in h5const.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_info_by_name()

This function implements the HDF5-API function H5Aget_info_by_name. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_info_by_name(
  attr_name,
  obj_name,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

attr_name

The name of the attribute.

obj_name

The name of the object, relative to the object the method is called on.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_info_by_idx()

This function implements the HDF5-API function H5Aget_info_by_idx. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_info_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

n

The index of the item to access; zero-based.

obj_name

The name of the object, relative to the object the method is called on.

idx_type

The index to iterate over; one of the H5_INDEX_* values in h5const.

order

The order in which to iterate; one of the H5_ITER_* values in h5const.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_name_by_idx()

This function implements the HDF5-API function H5Aget_name_by_idx. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_name_by_idx(
  n,
  obj_name,
  idx_type = h5const$H5_INDEX_NAME,
  order = h5const$H5_ITER_NATIVE,
  link_access_pl = h5const$H5P_DEFAULT
)

Arguments

n

The index of the item to access; zero-based.

obj_name

The name of the object, relative to the object the method is called on.

idx_type

The index to iterate over; one of the H5_INDEX_* values in h5const.

order

The order in which to iterate; one of the H5_ITER_* values in h5const.

link_access_pl

The link access property list. See H5P_LINK_ACCESS.

H5T$attr_get_number()

This function implements the HDF5-API function H5Aget_num_attrs. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_a.html for details.

Usage

H5T$attr_get_number()

H5T$create_reference()

This function implements the HDF5-API function H5Rcreate. If space=NULL then a H5R_OBJECT reference is created, otherwise a H5R_DATASET_REGION reference Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_r.html for details.

Usage

H5T$create_reference(name = ".", space = NULL)

Arguments

name

The name of the object, relative to the object the method is called on.

space

The dataspace to use; an object of class H5S.

H5T$new()

Internal use only

Usage

H5T$new(id)

Arguments

id

An HDF5 id; for internal use only.

H5T$get_class()

This function implements the HDF5-API function H5Tget_class. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$get_class()

H5T$get_size()

This function implements the HDF5-API function H5Tget_size. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$get_size(...)

Arguments

...

ignored

H5T$set_size()

This function implements the HDF5-API function H5Tset_size. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$set_size(size)

Arguments

size

The size of the object in bytes; use Inf for variable-length strings.

H5T$set_precision()

This function implements the HDF5-API function H5Tset_precision. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$set_precision(precision)

Arguments

precision

The precision of the datatype, in bits.

H5T$get_precision()

This function implements the HDF5-API function H5Tget_precision. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$get_precision()

H5T$set_order()

This function implements the HDF5-API function H5Tset_order. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$set_order(order)

Arguments

order

The order in which to iterate; one of the H5_ITER_* values in h5const.

H5T$get_order()

This function implements the HDF5-API function H5Tget_order. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$get_order()

H5T$set_offset()

This function implements the HDF5-API function H5Tset_offset. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$set_offset(offset)

Arguments

offset

The offset to apply; see the corresponding HDF5 function for how it is interpreted.

H5T$get_offset()

This function implements the HDF5-API function H5Tget_offset. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$get_offset()

H5T$set_pad()

This function implements the HDF5-API function H5Tset_pad. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$set_pad(pad)

Arguments

pad

The padding to use for unused bits; one of the H5T_PAD_* values in h5const.

H5T$get_pad()

This function implements the HDF5-API function H5Tget_pad. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$get_pad()

H5T$copy()

This function implements the HDF5-API function H5Tcopy. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$copy()

H5T$is_committed()

This function implements the HDF5-API function H5Tcommitted. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$is_committed()

H5T$equal()

This function implements the HDF5-API function H5Tequal. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$equal(dtype)

Arguments

dtype

The datatype to use; an object of class H5T.

H5T$is_vlen()

This function detects if the underlying type is H5T_VLEN or a variable length string. This is used to know if after reading a dataset, memory has to be freed

Usage

H5T$is_vlen()

H5T$detect_class()

This function implements the HDF5-API function H5Tdetect_class. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$detect_class(dtype_class)

Arguments

dtype_class

The datatype class to look for; one of the H5T_* class values in h5const.

H5T$get_native_type()

This function implements the HDF5-API function H5Tget_native_type. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$get_native_type(direction = h5const$H5T_DIR_ASCEND)

Arguments

direction

The direction in which to search for a native type; one of the H5T_DIR_* values in h5const.

H5T$get_create_plist()

This function implements the HDF5-API function H5Tget_create_plist. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T$get_create_plist()

H5T$to_text()

This function implements the HDF5-API function H5LTdtype_to_text.

Usage

H5T$to_text(lang_type = h5const$H5LT_DDL)

Arguments

lang_type

The language to produce the text representation in; one of the H5LT_LANG_* values in h5const.

H5T$print()

Prints information for the group

Usage

H5T$print(...)

Arguments

...

ignored

H5T$obj_info()

This function implements the HDF5-API function H5Oget_info. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_o.html for details.

Usage

H5T$obj_info(remove_internal_use_only = TRUE)

Arguments

remove_internal_use_only

Logical; should fields that are only of internal use be dropped from the result.

Examples

my_int <- h5types$H5T_NATIVE_INT
my_int$to_text()
#> [1] "H5T_STD_I32LE"
my_int$get_size()
#> [1] 4

# Show how to commit a datatype
fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
my_int$is_committed()
#> [1] FALSE
file$commit("my_int", my_int)
my_int$is_committed()
#> [1] TRUE

# can now also add attributes
h5attr(my_int, "test") <- "A string"
h5attributes(my_int)
#> $test
#> [1] "A string"
#> 

file$close_all()
file.remove(fname)
#> [1] TRUE