Class for HDF5 array datatypes.

Inherits from class H5T. This class represents an array. As datasets in HDF5 are itself already arrays, this datatype is not needed there. It is mostly useful when a column in a H5T_COMPUND object is intended to be an array. This however makes it difficult to work with such objects in R - as a column of the corresponding data.frame has to be an array. So please use with care.

Value

Object of class H5T_ARRAY.

See also

Author

Holger Hoefling

Super classes

H5RefClass -> H5T -> H5T_ARRAY

Methods

Public methods

H5T_ARRAY$new()

H5T_ARRAY$get_array_ndims()

H5T_ARRAY$get_array_dims()

H5T_ARRAY$get_super()

H5T_ARRAY$describe()

Inherited methods

H5RefClass$close()

H5RefClass$dec_ref()

H5RefClass$get_file_id()

H5RefClass$get_obj_type()

H5RefClass$get_ref()

H5RefClass$inc_ref()

H5RefClass$methods()

H5T$attr_delete()

H5T$attr_delete_by_idx()

H5T$attr_delete_by_name()

H5T$attr_exists()

H5T$attr_exists_by_name()

H5T$attr_get_number()

H5T$attr_info_by_idx()

H5T$attr_info_by_name()

H5T$attr_name_by_idx()

H5T$attr_open()

H5T$attr_open_by_idx()

H5T$attr_open_by_name()

H5T$attr_rename()

H5T$attr_rename_by_name()

H5T$copy()

H5T$create_attr()

H5T$create_attr_by_name()

H5T$create_reference()

H5T$detect_class()

H5T$equal()

H5T$get_class()

H5T$get_create_plist()

H5T$get_native_type()

H5T$get_obj_name()

H5T$get_offset()

H5T$get_order()

H5T$get_pad()

H5T$get_precision()

H5T$get_size()

H5T$is_committed()

H5T$is_vlen()

H5T$obj_info()

H5T$print()

H5T$set_offset()

H5T$set_order()

H5T$set_pad()

H5T$set_precision()

H5T$set_size()

H5T$to_text()

H5T_ARRAY$new()

Create an array datatype.

Usage

H5T_ARRAY$new(dims, dtype_base, id = NULL)

Arguments

dims

The dimension of the datatype

dtype_base

The datatype that makes up the elements of the array

id

internal use only

H5T_ARRAY$get_array_ndims()

This function implements the HDF5-API function H5Tget_array_ndims. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_ARRAY$get_array_ndims()

H5T_ARRAY$get_array_dims()

This function implements the HDF5-API function H5Tget_array_dims2. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_ARRAY$get_array_dims()

H5T_ARRAY$get_super()

This function implements the HDF5-API function H5Tget_super. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_ARRAY$get_super()

H5T_ARRAY$describe()

Print a detailed description of the datatype; this is experimental

Usage

H5T_ARRAY$describe()