Class for HDF5 floating point datatypes.

Inherits from class H5T. Users should not create float types with this class, but instead use e.g. h5types$H5T_NATIVE_DOUBLE. Using the functions of this class, many aspects of the representation of the floating point number can then be manipulated.

Value

Object of class H5T_FLOAT.

See also

Author

Holger Hoefling

Super classes

H5RefClass -> H5T -> H5T_FLOAT

Methods

Public methods

H5T_FLOAT$set_fields()

H5T_FLOAT$get_fields()

H5T_FLOAT$set_ebias()

H5T_FLOAT$get_ebias()

H5T_FLOAT$set_norm()

H5T_FLOAT$get_norm()

H5T_FLOAT$set_inpad()

H5T_FLOAT$get_inpad()

H5T_FLOAT$describe()

Inherited methods

H5RefClass$close()

H5RefClass$dec_ref()

H5RefClass$get_file_id()

H5RefClass$get_obj_type()

H5RefClass$get_ref()

H5RefClass$inc_ref()

H5RefClass$methods()

H5T$attr_delete()

H5T$attr_delete_by_idx()

H5T$attr_delete_by_name()

H5T$attr_exists()

H5T$attr_exists_by_name()

H5T$attr_get_number()

H5T$attr_info_by_idx()

H5T$attr_info_by_name()

H5T$attr_name_by_idx()

H5T$attr_open()

H5T$attr_open_by_idx()

H5T$attr_open_by_name()

H5T$attr_rename()

H5T$attr_rename_by_name()

H5T$copy()

H5T$create_attr()

H5T$create_attr_by_name()

H5T$create_reference()

H5T$detect_class()

H5T$equal()

H5T$get_class()

H5T$get_create_plist()

H5T$get_native_type()

H5T$get_obj_name()

H5T$get_offset()

H5T$get_order()

H5T$get_pad()

H5T$get_precision()

H5T$get_size()

H5T$initialize()

H5T$is_committed()

H5T$is_vlen()

H5T$obj_info()

H5T$print()

H5T$set_offset()

H5T$set_order()

H5T$set_pad()

H5T$set_precision()

H5T$set_size()

H5T$to_text()

H5T_FLOAT$set_fields()

This function implements the HDF5-API function H5Tset_fields. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_FLOAT$set_fields(spos, epos, esize, mpos, msize)

Arguments

spos

The bit position of the sign bit.

epos

The bit position of the exponent field.

esize

The size of the exponent field, in bits.

mpos

The bit position of the mantissa field.

msize

The size of the mantissa field, in bits.

H5T_FLOAT$get_fields()

This function implements the HDF5-API function H5Tget_fields. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_FLOAT$get_fields()

H5T_FLOAT$set_ebias()

This function implements the HDF5-API function H5Tset_ebias. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_FLOAT$set_ebias(ebias)

Arguments

ebias

The exponent bias.

H5T_FLOAT$get_ebias()

This function implements the HDF5-API function H5Tget_ebias. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_FLOAT$get_ebias()

H5T_FLOAT$set_norm()

This function implements the HDF5-API function H5Tset_norm. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_FLOAT$set_norm(norm)

Arguments

norm

The mantissa normalisation; one of the H5T_NORM_* values in h5const.

H5T_FLOAT$get_norm()

This function implements the HDF5-API function H5Tget_norm. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_FLOAT$get_norm()

H5T_FLOAT$set_inpad()

This function implements the HDF5-API function H5Tset_inpad. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_FLOAT$set_inpad(inpad)

Arguments

inpad

How to fill unused internal bits; one of the H5T_PAD_* values in h5const.

H5T_FLOAT$get_inpad()

This function implements the HDF5-API function H5Tget_inpad. Please see the documentation at https://support.hdfgroup.org/documentation/hdf5/latest/group___h5_t.html for details.

Usage

H5T_FLOAT$get_inpad()

H5T_FLOAT$describe()

Print a detailed description of the datatype; this is experimental

Usage

H5T_FLOAT$describe()