Class for HDF5 floating point datatypes.

Inherits from class H5T. Users should not create float types with this class, but instead use e.g. h5types$H5T_NATIVE_DOUBLE. Using the functions of this class, many aspects of the representation of the floating point number can then be manipulated.

Value

Object of class H5T_FLOAT.

Methods

set_fields(spos, epos, esize, mpos, msize)

This function implements the HDF5-API function H5Tset_fields. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5T.html#Datatype-SetFields for details.

get_fields()

This function implements the HDF5-API function H5Tget_fields. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5T.html#Datatype-GetFields for details.

set_ebias(ebias)

This function implements the HDF5-API function H5Tset_ebias. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5T.html#Datatype-SetEbias for details.

get_ebias()

This function implements the HDF5-API function H5Tget_ebias. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5T.html#Datatype-GetEbias for details.

set_norm(norm)

This function implements the HDF5-API function H5Tset_norm. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5T.html#Datatype-SetNorm for details.

get_norm()

This function implements the HDF5-API function H5Tget_norm. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5T.html#Datatype-GetNorm for details.

set_inpad(inpad)

This function implements the HDF5-API function H5Tset_inpad. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5T.html#Datatype-SetInpad for details.

get_inpad()

This function implements the HDF5-API function H5Tget_inpad. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5T.html#Datatype-GetInpad for details.

describe()

Print a detailed description of the datatype; this is experimental

See also

H5T