Various functions for factor_ext objects

Various functions for factor_ext objects

values(x, ...)

# S3 method for factor_ext
values(x, ...)

# S3 method for factor
values(x, ...)

# S3 method for default
values(x, ...)

# S3 method for factor_ext
as.character(x, ...)

# S3 method for factor_ext
[[(x, ...)

# S3 method for factor_ext
[[(x, ...) <- value

# S3 method for factor_ext
[(x, ..., drop = FALSE)

# S3 method for factor_ext
[(x, ...) <- value

is.factor_ext(x)

coercible_to_factor(x)

coerce_to_factor(x)

# S3 method for factor_ext
print(x, quote = FALSE, max.levels = NULL,
  width = getOption("width"), ...)

# S3 method for factor_ext
==(e1, e2)

# S3 method for factor_ext
!=(e1, e2)

# S3 method for factor_ext
c(...)

Arguments

x
Object of type factor_ext

...
Currently ignored

value
The object to assign; here has be a level of factor_ext

drop
Should dimensions of size 1 be dropped?

quote
logical, indicating whether or not strings should be printed with surrounding quotes.

max.levels
integer, indicating how many levels should be printed. if '0', no extra "Levels" line will be printed. The default, 'NULL', entails choosing 'max.levels' such that the levels print on one line of width 'width' (same for values).

width
only used when max.levels is NULL (see above)

e1, e2
The two objects in the equality or inequality comparison.

Value

Depending on the function

Details

values

Extracts the underlying values of an object (the generic here)

values.factor_ext

Extracts the underlying values of a factor_ext object

values.factor

Extracts the underlying values of a factor

values.default

Default of the values function; currently returns an error

as.character

Coerces factor_ext to a character-representation using it levels, not values

[[.factor_ext

Single-item subsetting of a factor_ext object

[[<-.factor_ext

Single-item subset assignment to a factor_ext object

[.factor_ext

Subsetting of a factor_ext object

[<-.factor_ext

Subset assignment to a factor_ext object

is.factor_ext

Check if it is a factor_ext object. Returns a logical

coercible_to_factor

Checks if a factor_ext could be coerced to a factor. Return a logical.

coerce_to_factor

Coerces to a factor, otherwise throws an error if not possible.

print.factor_ext

Prints a factor_ext object.

==.factor_ext

Compare two factor_ext for equality.

!=.factor_ext

Compare two factor_ext for inequality.

c.factor_ext

Concatenate objects of type factor_ext.