Wrapper functions to provide an h5 compatible interface.

The functions listed below provide a wrapper-interface compatible to functions specified in the h5 package. The author(s) have decided to deprecate h5 and join forces and still make the transition for h5 users as smooth as possible. Additionally, almost all testcases could be transferred to hdf5r to improve test coverage even more.

h5file(...)

createGroup(object, name, ...)

openLocation(object, name)

openGroup(object, name)

createDataSet(object, name, ...)

readDataSet(object)

h5close(object)

h5flush(object)

existsGroup(object, name)

is.h5file(name)

extendDataSet(object, dims)

# S3 method for H5D
rbind(x, mat, ..., deparse.level = 1)

# S3 method for H5D
cbind(x, mat, ..., deparse.level = 1)

# S3 method for H5D
c(x, ...)

h5unlink(object, name)

list.attributes(object)

Arguments

...
Additional parameters passed to create_group or h5file.

object
CommonFG; Object implementing the CommonFG Interface (e.g. H5File, H5Group).

name
Name of the group to create.

dims
numeric; Dimension vector to which dataset should be extended.

x
An object of class H5D; the dataset to add rows or columns to; Needs to be a matrix

mat
The matrix to add to x

deparse.level
Set to 1; ignored otherwise; only present as required by generic

Details

Below you can find a list of all h5 functions including hdf5r mappings.

h5file

Directly maps to H5File$new, see also H5File.

createGroup

Maps to object$create_group where object implements CommonFG.

openLocation

Uses object$open where object implements CommonFG.

createDataSet

Maps to object$create_dataset where object implements CommonFG.

readDataSet

Maps to object$read, see also H5D.

h5close

Maps to object$close_all for H5File and object$close for other.

h5flush

Maps to object$flush where object implements CommonFGDTA.

The following interfaces are defined:

CommonFG

Implemented by objects of class H5File and H5Group.

CommonFGDTA

Implemented by objects of class H5File, H5Group, H5D, H5T and H5A.

References

Mario Annau (2017). h5: Interface to the 'HDF5' Library. R package version 0.9.9. https://github.com/mannau/h5