These are all types that are used in HDF5

HDF5 provides many native datatypes. These are all stored in the h5types environment. An overview of all available types can be seen using h5types$overview. Any specific type can be accessed using the $-operator. See also the examples below.

Examples

h5types$overview

#>        Category                             Name            Type_id
#> 1          IEEE                   H5T_IEEE_F32BE 216172782113783863
#> 2          IEEE                   H5T_IEEE_F32LE 216172782113783862
#> 3          IEEE                   H5T_IEEE_F64BE 216172782113783865
#> 4          IEEE                   H5T_IEEE_F64LE 216172782113783864
#> 5           STD                     H5T_STD_I8BE 216172782113783869
#> 6           STD                     H5T_STD_I8LE 216172782113783868
#> 7           STD                    H5T_STD_I16BE 216172782113783871
#> 8           STD                    H5T_STD_I16LE 216172782113783870
#> 9           STD                    H5T_STD_I32BE 216172782113783873
#> 10          STD                    H5T_STD_I32LE 216172782113783872
#> 11          STD                    H5T_STD_I64BE 216172782113783875
#> 12          STD                    H5T_STD_I64LE 216172782113783874
#> 13          STD                     H5T_STD_U8BE 216172782113783877
#> 14          STD                     H5T_STD_U8LE 216172782113783876
#> 15          STD                    H5T_STD_U16BE 216172782113783879
#> 16          STD                    H5T_STD_U16LE 216172782113783878
#> 17          STD                    H5T_STD_U32BE 216172782113783881
#> 18          STD                    H5T_STD_U32LE 216172782113783880
#> 19          STD                    H5T_STD_U64BE 216172782113783883
#> 20          STD                    H5T_STD_U64LE 216172782113783882
#> 21          STD                     H5T_STD_B8BE 216172782113783885
#> 22          STD                     H5T_STD_B8LE 216172782113783884
#> 23          STD                    H5T_STD_B16BE 216172782113783887
#> 24          STD                    H5T_STD_B16LE 216172782113783886
#> 25          STD                    H5T_STD_B32BE 216172782113783889
#> 26          STD                    H5T_STD_B32LE 216172782113783888
#> 27          STD                    H5T_STD_B64BE 216172782113783891
#> 28          STD                    H5T_STD_B64LE 216172782113783890
#> 29          STD                  H5T_STD_REF_OBJ 216172782113783899
#> 30          STD              H5T_STD_REF_DSETREG 216172782113783900
#> 31          STD                      H5T_STD_REF 216172782113783901
#> 32         UNIX                   H5T_UNIX_D32BE 216172782113783893
#> 33         UNIX                   H5T_UNIX_D32LE 216172782113783892
#> 34         UNIX                   H5T_UNIX_D64BE 216172782113783895
#> 35         UNIX                   H5T_UNIX_D64LE 216172782113783894
#> 36       STRING                         H5T_C_S1 216172782113783897
#> 37       STRING                   H5T_FORTRAN_S1 216172782113783898
#> 38        INTEL                     H5T_INTEL_I8 216172782113783868
#> 39        INTEL                    H5T_INTEL_I16 216172782113783870
#> 40        INTEL                    H5T_INTEL_I32 216172782113783872
#> 41        INTEL                    H5T_INTEL_I64 216172782113783874
#> 42        INTEL                     H5T_INTEL_U8 216172782113783876
#> 43        INTEL                    H5T_INTEL_U16 216172782113783878
#> 44        INTEL                    H5T_INTEL_U32 216172782113783880
#> 45        INTEL                    H5T_INTEL_U64 216172782113783882
#> 46        INTEL                     H5T_INTEL_B8 216172782113783884
#> 47        INTEL                    H5T_INTEL_B16 216172782113783886
#> 48        INTEL                    H5T_INTEL_B32 216172782113783888
#> 49        INTEL                    H5T_INTEL_B64 216172782113783890
#> 50        INTEL                    H5T_INTEL_F32 216172782113783862
#> 51        INTEL                    H5T_INTEL_F64 216172782113783864
#> 52        ALPHA                     H5T_ALPHA_I8 216172782113783868
#> 53        ALPHA                    H5T_ALPHA_I16 216172782113783870
#> 54        ALPHA                    H5T_ALPHA_I32 216172782113783872
#> 55        ALPHA                    H5T_ALPHA_I64 216172782113783874
#> 56        ALPHA                     H5T_ALPHA_U8 216172782113783876
#> 57        ALPHA                    H5T_ALPHA_U16 216172782113783878
#> 58        ALPHA                    H5T_ALPHA_U32 216172782113783880
#> 59        ALPHA                    H5T_ALPHA_U64 216172782113783882
#> 60        ALPHA                     H5T_ALPHA_B8 216172782113783884
#> 61        ALPHA                    H5T_ALPHA_B16 216172782113783886
#> 62        ALPHA                    H5T_ALPHA_B32 216172782113783888
#> 63        ALPHA                    H5T_ALPHA_B64 216172782113783890
#> 64        ALPHA                    H5T_ALPHA_F32 216172782113783862
#> 65        ALPHA                    H5T_ALPHA_F64 216172782113783864
#> 66         MIPS                      H5T_MIPS_I8 216172782113783869
#> 67         MIPS                     H5T_MIPS_I16 216172782113783871
#> 68         MIPS                     H5T_MIPS_I32 216172782113783873
#> 69         MIPS                     H5T_MIPS_I64 216172782113783875
#> 70         MIPS                      H5T_MIPS_U8 216172782113783877
#> 71         MIPS                     H5T_MIPS_U16 216172782113783879
#> 72         MIPS                     H5T_MIPS_U32 216172782113783881
#> 73         MIPS                     H5T_MIPS_U64 216172782113783883
#> 74         MIPS                      H5T_MIPS_B8 216172782113783885
#> 75         MIPS                     H5T_MIPS_B16 216172782113783887
#> 76         MIPS                     H5T_MIPS_B32 216172782113783889
#> 77         MIPS                     H5T_MIPS_B64 216172782113783891
#> 78         MIPS                     H5T_MIPS_F32 216172782113783863
#> 79         MIPS                     H5T_MIPS_F64 216172782113783865
#> 80          VAX                      H5T_VAX_F32 216172782113783866
#> 81          VAX                      H5T_VAX_F64 216172782113783867
#> 82       NATIVE                  H5T_NATIVE_CHAR 216172782113783816
#> 83       NATIVE                 H5T_NATIVE_SCHAR 216172782113783816
#> 84       NATIVE                 H5T_NATIVE_UCHAR 216172782113783817
#> 85       NATIVE                 H5T_NATIVE_SHORT 216172782113783818
#> 86       NATIVE                H5T_NATIVE_USHORT 216172782113783819
#> 87       NATIVE                   H5T_NATIVE_INT 216172782113783820
#> 88       NATIVE                  H5T_NATIVE_UINT 216172782113783821
#> 89       NATIVE                  H5T_NATIVE_LONG 216172782113783822
#> 90       NATIVE                 H5T_NATIVE_ULONG 216172782113783823
#> 91       NATIVE                 H5T_NATIVE_LLONG 216172782113783848
#> 92       NATIVE                H5T_NATIVE_ULLONG 216172782113783849
#> 93       NATIVE                 H5T_NATIVE_FLOAT 216172782113783850
#> 94       NATIVE                H5T_NATIVE_DOUBLE 216172782113783851
#> 95       NATIVE                    H5T_NATIVE_B8 216172782113783853
#> 96       NATIVE                   H5T_NATIVE_B16 216172782113783854
#> 97       NATIVE                   H5T_NATIVE_B32 216172782113783855
#> 98       NATIVE                   H5T_NATIVE_B64 216172782113783856
#> 99       NATIVE                H5T_NATIVE_OPAQUE 216172782113783896
#> 100      NATIVE                 H5T_NATIVE_HADDR 216172782113783857
#> 101      NATIVE                 H5T_NATIVE_HSIZE 216172782113783858
#> 102      NATIVE                H5T_NATIVE_HSSIZE 216172782113783859
#> 103      NATIVE                  H5T_NATIVE_HERR 216172782113783860
#> 104      NATIVE                 H5T_NATIVE_HBOOL 216172782113783861
#> 105      NATIVE                  H5T_NATIVE_INT8 216172782113783824
#> 106      NATIVE                 H5T_NATIVE_UINT8 216172782113783825
#> 107      NATIVE            H5T_NATIVE_INT_LEAST8 216172782113783826
#> 108      NATIVE           H5T_NATIVE_UINT_LEAST8 216172782113783827
#> 109      NATIVE             H5T_NATIVE_INT_FAST8 216172782113783828
#> 110      NATIVE            H5T_NATIVE_UINT_FAST8 216172782113783829
#> 111      NATIVE                 H5T_NATIVE_INT16 216172782113783830
#> 112      NATIVE                H5T_NATIVE_UINT16 216172782113783831
#> 113      NATIVE           H5T_NATIVE_INT_LEAST16 216172782113783832
#> 114      NATIVE          H5T_NATIVE_UINT_LEAST16 216172782113783833
#> 115      NATIVE            H5T_NATIVE_INT_FAST16 216172782113783834
#> 116      NATIVE           H5T_NATIVE_UINT_FAST16 216172782113783835
#> 117      NATIVE                 H5T_NATIVE_INT32 216172782113783836
#> 118      NATIVE                H5T_NATIVE_UINT32 216172782113783837
#> 119      NATIVE           H5T_NATIVE_INT_LEAST32 216172782113783838
#> 120      NATIVE          H5T_NATIVE_UINT_LEAST32 216172782113783839
#> 121      NATIVE            H5T_NATIVE_INT_FAST32 216172782113783840
#> 122      NATIVE           H5T_NATIVE_UINT_FAST32 216172782113783841
#> 123      NATIVE                 H5T_NATIVE_INT64 216172782113783842
#> 124      NATIVE                H5T_NATIVE_UINT64 216172782113783843
#> 125      NATIVE           H5T_NATIVE_INT_LEAST64 216172782113783844
#> 126      NATIVE          H5T_NATIVE_UINT_LEAST64 216172782113783845
#> 127      NATIVE            H5T_NATIVE_INT_FAST64 216172782113783846
#> 128      NATIVE           H5T_NATIVE_UINT_FAST64 216172782113783847
#> 129 C_API_types                      H5E_major_t 216172782113783842
#> 130 C_API_types                      H5E_minor_t 216172782113783842
#> 131 C_API_types                       H5FD_mem_t 216172782113783836
#> 132 C_API_types                       H5G_link_t 216172782113783836
#> 133 C_API_types                H5O_msg_crt_idx_t 216172782113783837
#> 134 C_API_types                     H5Z_filter_t 216172782113783836
#> 135 C_API_types                          haddr_t 216172782113783843
#> 136 C_API_types                          hbool_t 216172782113783825
#> 137 C_API_types                           herr_t 216172782113783836
#> 138 C_API_types                            hid_t 216172782113783842
#> 139 C_API_types                          hsize_t 216172782113783843
#> 140 C_API_types                         hssize_t 216172782113783842
#> 141 C_API_types                           htri_t 216172782113783836
#> 142 C_API_types                              int 216172782113783836
#> 143 C_API_types                          int64_t 216172782113783842
#> 144 C_API_types                             long 216172782113783842
#> 145 C_API_types                         long_int 216172782113783842
#> 146 C_API_types                        long_long 216172782113783842
#> 147 C_API_types                            off_t 216172782113783842
#> 148 C_API_types                            short 216172782113783830
#> 149 C_API_types                           size_t 216172782113783843
#> 150 C_API_types                          ssize_t 216172782113783842
#> 151 C_API_types                           time_t 216172782113783842
#> 152 C_API_types                         uint32_t 216172782113783837
#> 153 C_API_types                         uint64_t 216172782113783843
#> 154 C_API_types                          uint8_t 216172782113783825
#> 155 C_API_types                         unsigned 216172782113783837
#> 156 C_API_types                    unsigned_char 216172782113783825
#> 157 C_API_types                     unsigned_int 216172782113783837
#> 158 C_API_types                    unsigned_long 216172782113783843
#> 159 C_API_types               unsigned_long_long 216172782113783843
#> 160 C_API_types                   unsigned_short 216172782113783831
#> 161 C_API_types                             char 216172782113784122
#> 162 C_API_types                           double 216172782113783851
#> 163 C_API_types                            float 216172782113783850
#> 164        enum                       H5_index_t 216172782113784157
#> 165        enum                  H5_iter_order_t 216172782113784156
#> 166        enum              H5C_cache_decr_mode 216172782113784123
#> 167        enum        H5C_cache_flash_incr_mode 216172782113784124
#> 168        enum              H5C_cache_incr_mode 216172782113784125
#> 169        enum                 H5D_alloc_time_t 216172782113784126
#> 170        enum                H5D_chunk_index_t 216172782113784127
#> 171        enum                  H5D_fill_time_t 216172782113784128
#> 172        enum                 H5D_fill_value_t 216172782113784129
#> 173        enum                     H5D_layout_t 216172782113784130
#> 174        enum H5D_mpio_actual_chunk_opt_mode_t 216172782113784153
#> 175        enum        H5D_mpio_actual_io_mode_t 216172782113784154
#> 176        enum   H5D_mpio_no_collective_cause_t 216172782113784155
#> 177        enum               H5D_space_status_t 216172782113784131
#> 178        enum                   H5D_vds_view_t 216172782113784132
#> 179        enum                  H5E_direction_t 216172782113784133
#> 180        enum                       H5E_type_t 216172782113784134
#> 181        enum                    H5ES_status_t 216172782113784135
#> 182        enum               H5F_close_degree_t 216172782113784137
#> 183        enum            H5F_file_space_type_t 216172782113784138
#> 184        enum            H5F_fspace_strategy_t 216172782113784139
#> 185        enum                     H5F_libver_t 216172782113784140
#> 186        enum                        H5F_mem_t 216172782113784141
#> 187        enum                      H5F_scope_t 216172782113784142
#> 188        enum             H5FD_file_image_op_t 216172782113784136
#> 189        enum                        H5G_obj_t 216172782113784143
#> 190        enum               H5G_storage_type_t 216172782113784144
#> 191        enum                       H5I_type_t 216172782113784145
#> 192        enum                       H5L_type_t 216172782113784146
#> 193        enum                      H5LT_lang_t 216172782113784147
#> 194        enum            H5O_mcdt_search_ret_t 216172782113784150
#> 195        enum                       H5O_type_t 216172782113784151
#> 196        enum                      H5PL_type_t 216172782113784152
#> 197        enum                       H5R_type_t 216172782113784158
#> 198        enum                      H5S_class_t 216172782113784160
#> 199        enum                     H5S_sel_type 216172782113784159
#> 200        enum                    H5S_seloper_t 216172782113784161
#> 201        enum                        H5T_bkg_t 216172782113784162
#> 202        enum                      H5T_class_t 216172782113784163
#> 203        enum                        H5T_cmd_t 216172782113784164
#> 204        enum                H5T_conv_except_t 216172782113784165
#> 205        enum                   H5T_conv_ret_t 216172782113784166
#> 206        enum                       H5T_cset_t 216172782113784167
#> 207        enum                  H5T_direction_t 216172782113784168
#> 208        enum                       H5T_norm_t 216172782113784169
#> 209        enum                      H5T_order_t 216172782113784170
#> 210        enum                        H5T_pad_t 216172782113784171
#> 211        enum                       H5T_pers_t 216172782113784172
#> 212        enum                       H5T_sign_t 216172782113784173
#> 213        enum                        H5T_str_t 216172782113784174
#> 214        enum                   H5VL_map_get_t 216172782113784148
#> 215        enum              H5VL_map_specific_t 216172782113784149
#> 216        enum                  H5Z_cb_return_t 216172782113784175
#> 217        enum                        H5Z_EDC_t 216172782113784176
#> 218        enum              H5Z_SO_scale_type_t 216172782113784177
#> 219      struct               H5L_info1_helper_t 216172782113784186
#> 220      struct               H5L_info2_helper_t 216172782113784203
#> 221      struct                     H5_ih_info_t 216172782113784178
#> 222      struct                       H5A_info_t 216172782113784179
#> 223      struct              H5AC_cache_config_t 216172782113784185
#> 224      struct        H5AC_cache_image_config_t 216172782113784182
#> 225      struct                       H5D_info_t 216172782113784183
#> 226      struct                     H5E_error1_t 216172782113784191
#> 227      struct                     H5E_error2_t 216172782113784192
#> 228      struct               H5F_info1_helper_t 216172782113784197
#> 229      struct                      H5F_info1_t 216172782113784205
#> 230      struct                 H5F_info2_free_t 216172782113784193
#> 231      struct                 H5F_info2_sohm_t 216172782113784198
#> 232      struct                H5F_info2_super_t 216172782113784194
#> 233      struct                      H5F_info2_t 216172782113784206
#> 234      struct                 H5F_retry_info_t 216172782113784195
#> 235      struct                  H5F_sect_info_t 216172782113784180
#> 236      struct                       H5G_info_t 216172782113784184
#> 237      struct                      H5L_info1_t 216172782113784200
#> 238      struct                      H5L_info2_t 216172782113784210
#> 239      struct                      H5ls_info_t 216172782113784207
#> 240      struct        H5O_hdr_info_helper_msg_t 216172782113784187
#> 241      struct      H5O_hdr_info_helper_space_t 216172782113784181
#> 242      struct                   H5O_hdr_info_t 216172782113784201
#> 243      struct                H5O_info_helper_t 216172782113784199
#> 244      struct                      H5O_info1_t 216172782113784208
#> 245      struct                      H5O_info2_t 216172782113784204
#> 246      struct                H5O_native_info_t 216172782113784209
#> 247      struct                       H5O_stat_t 216172782113784196
#> 248      struct                      H5O_token_t 216172782113784189
#> 249      struct                 H5R_ref_helper_t 216172782113784188
#> 250      struct                        H5R_ref_t 216172782113784202
#> 251      struct                  hdset_reg_ref_t 216172782113784190
#> 252      NATIVE               H5T_NATIVE_LDOUBLE 216172782113783852
#> 253      Custom                      H5T_LOGICAL 216172782113784618
#> 254      Custom                   H5T_LOGICAL_NA 216172782113784627

h5types$H5T_NATIVE_INT

#> Class: H5T_INTEGER
#> Datatype: H5T_STD_I32LE

h5types$H5T_NATIVE_DOUBLE

#> Class: H5T_FLOAT
#> Datatype: H5T_IEEE_F64LE