Single hyperslab dimension to explicit vector

Single hyperslab dimension to explicit vector

hyperslab_to_points(hyperslab)

Arguments

hyperslab
a length 4 vector describing the start, count, stride and block component of a single dimension of a hyperslab

Value

An explicit vector describing the points in the hyperslab dimension

Details

Uses the information of a hyperslab and turns it into an explicit vector.