Print a data frame with extended factor objects

Print a data frame that includes extended factor objects

# S3 method for data.frame_ext
print(x, ...)

Arguments

x
The data.frame_ext object to print; Is returned by ls from H5File and H5Group and this function allows for petter printing of factor_ext so that the label instead of the value is printed.

...
Parameters to be passed on directly to print.data.frame

Value

The object to print itself, invisibly

Details

The regular print function for data-frames has special methods built-in for factors so that the label is printed instead of the constant. This function is intended to provide the same functionality for data frames with extended factors, by adding the class data.frame_ext to the class vector.