Convert a text description to a datatype

Convert a text description to a datatype

text_to_dtype(text, lang_type = h5const$H5LT_DDL)

Arguments

text
The text to convert to the datatype

lang_type
The type of language to use; currently only H5LT_DDL is supported.

Value

A datatype corresponding to the text with the appropriate class inheriting from H5T.

Details

Converts a text to a datatype using the HDF5 function H5LT_text_to_dtype. Documentation can be found at https://www.hdfgroup.org/HDF5/doc/HL/RM_H5LT.html#H5LTtext_to_dtype.