Class for HDF5 property lists.

This is the base class for all property lists, but most have a specialised class. It inherits all functions of the H5RefClass-class. It is also the base class for many other classes, specifically

Dataset Creation

H5P_DATASET_CREATE-class

Dataset Access

H5P_DATASET_ACCESS-class

Dataset Transfer

H5P_DATASET_XFER-class

Link Creation

H5P_LINK_CREATE-class

Link Access

H5P_LINK_ACCESS-class

Object Creation

H5P_OBJECT_CREATE-class

Object Copy

H5P_OBJECT_COPY-class

Attribute Creation

H5P_ATTRIBUTE_CREATE-class

The base class is unlikely to be needed by users - they should use the appropriate subclass required.

Value

Object of class H5P.

Methods

new(id = NULL)

Create a new property list; this function itself is unlikely to be needed by users. Users should use the classes of the type they actually require Parameters

id

Internal use only

get_class()

This function implements the HDF5-API function H5Pget_class. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-GetClass for details.

get_class_name()

This function implements the HDF5-API function H5Pget_class_name. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-GetClassName for details.

copy()

This function implements the HDF5-API function H5Pcopy. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-Copy for details.

equal(cmp)

This function implements the HDF5-API function H5Pequal. Please see the documentation at https://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html#Property-Equal for details.