Class for HDF5 property lists.

This is the base class for all property lists, but most have a specialized class. It inherits all functions of the H5RefClass. It is also the base class for many other classes, specifically

Dataset Creation

H5P_DATASET_CREATE

Dataset Access

H5P_DATASET_ACCESS

Dataset Transfer

H5P_DATASET_XFER

Link Creation

H5P_LINK_CREATE

Link Access

H5P_LINK_ACCESS

Object Creation

H5P_OBJECT_CREATE

Object Copy

H5P_OBJECT_COPY

Attribute Creation

H5P_ATTRIBUTE_CREATE

The base class is unlikely to be needed by users - they should use the appropriate subclass required.

Value

Object of class H5P.

Methods

new(id = NULL)

Create a new property list; this function itself is unlikely to be needed by users. Users should use the classes of the type they actually require

Parameters

id

Internal use only

get_class()

This function implements the HDF5-API function H5Pget_class. Please see the documentation at https://docs.hdfgroup.org/hdf5/develop/group___h5_p.html for details.

get_class_name()

This function implements the HDF5-API function H5Pget_class_name. Please see the documentation at https://docs.hdfgroup.org/hdf5/develop/group___h5_p.html for details.

copy()

This function implements the HDF5-API function H5Pcopy. Please see the documentation at https://docs.hdfgroup.org/hdf5/develop/group___h5_p.html for details.

equal(cmp)

This function implements the HDF5-API function H5Pequal. Please see the documentation at https://docs.hdfgroup.org/hdf5/develop/group___h5_p.html for details.

Author

Holger Hoefling

Examples

fname <- tempfile(fileext = ".h5")
file <- H5File$new(fname, mode = "a")
file[["testdataset"]] <- 1:10
p <- file[["testdataset"]]$get_create_plist()
p$get_class()
#> Class: H5P_CLASS
p$get_class_name()
#> [1] "dataset create"
p$copy()
#> Class: H5P_DATASET_CREATE
p$equal(p)
#> [1] TRUE
file$close_all()