Various functions for H5R objects

Various functions for H5R objects

is.H5R(x)

is.H5R_OBJECT(x)

is.H5R_DATASET_REGION(x)

# S3 method for H5R
names(x)

# S3 method for H5R
length(x)

# S3 method for H5R
[(x, i, j, ..., drop = TRUE)

# S3 method for H5R
[(x, i, ...) <- value

# S3 method for H5R
c(..., recursive = FALSE)

# S3 method for H5R
dim(x)

# S3 method for H5R
dim(x) <- value

# S3 method for H5R
t(x)

# S3 method for H5R
dimnames(x)

# S3 method for H5R
dimnames(x) <- value

# S3 method for H5R
cbind(..., deparse.level = 1)

# S3 method for H5R
rbind(..., deparse.level = 1)

# S3 method for H5R
print(x, ...)

# S3 method for H5R
format(x, ...)

# S3 method for H5R
as.data.frame(x, row.names = NULL, optional = FALSE, ...,
  nm = paste(deparse(substitute(x), width.cutoff = 500L), collapse =
  " "))

# S3 method for H5R
as.vector(x, mode = "any")

Arguments

x
Object of type H5R

i
First dimension

j
Second dimension

...
Any other dimensions (for subsetting), or objects to concatenate (for c) or combine by row/col (for cbind or rbind) or ignored (for print and format)

drop
Should dimensions of size 1 be dropped; LOGICAL

value
The value in an assignment

recursive
Ignored here

deparse.level
integer controlling the construction of labels in the case of non-matrix-like arguments (for the default method): 'deparse.level = 0' constructs no labels; the default, 'deparse.level = 1' constructs labels from the argument names

row.names
NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional
logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional.

nm
The column names to use

mode
Only 'any' supported

width.cutoff
ignored

collapse
ignored

Value

Depending on the function

Details

is.H5R

Check if object inherits from H5R

is.H5R_OBJECT

Check if object inherits from H5R_OBJECT

is.H5R_DATASET_REGION

Check if object inherits from H5R_DATASET_REGION

names.H5R

Returns the names of the elements of the vector

length.H5R

Returns the length of the vector

[.H5R

Array subsetting function

[<-.H5R

Array subset assignment

c.H5R

Concatenation of H5R vectors

dim.H5R

Dimensionality of the object

dim<-.H5R

Assign dimension of the object

t.H5R

Transpose a matrix of H5R objects

dimnames.H5R

Get the dimnames of the object

dimnames<-.H5R

Set the dimnames of the object

cbind.H5R

cbind functionality for H5R objects

rbind.H5R

rbind functionality for H5R objects

print.H5R

Printing of an object of class h5R

format.H5R

Formatting of an H5R object

as.data.frame.H5R

Coerce an H5R object to a data.frame

as.vector.H5R

Coerce to a vector

as.data.frame.H5R

Coerces the object to a data.frame

as.vector.H5R

Coerces to a vector