All constants used in HDF5

These are all constants used in HDF5. They are stored in an environment with locked bindings so that they cannot be changed. An overview of all constants can be seen with h5const$overview, listing all of them. Each constant can be accessed using $ and the name of the constant. See the examples below.

Details

There are also some flags that govern edge cases of conversion from HDF5 to R. This is related to how integers are being treated and the issue of R not being able to natively represent 64bit integers and not at all being able to represent unsigned 64bit integers (even using add-on packages). The constants all start with the term H5TOR. There are currently possible values

H5TOR_CONV_NONE

Doesn't do any conversion. Every integer datatype with more than 32 bit is returned as 64bit integers. For unsigned 64bit integers, the conversion to signed 64bit integers is done by truncation

H5TOR_CONV_INT64_INT_NOLOSS

Under this setting, whenever a 64 bit integer would be returned, it is checked if it would also fit into a 32 bit integer without data loss and returned as such if possible

H5TOR_CONV_INT64_FLOAT_NOLOSS

Under this setting, whenever a 64 bit integer would be returned, it is checked if it would also fit into a 64 bit floating point value without data loss and returned as such if possible

H5TOR_CONV_INT64_NOLOSS

Combines H5TOR_CONV_INT64_INT_NOLOSS and H5TOR_CONV_INT64_FLOAT_NOLOSS and is set as the default in the hdf5r.h5tor_default option.

H5TOR_CONV_INT64_FLOAT_FORCE

Under this setting, whenever a 64 bit integer would be returned, it is coerced to a double even if this results in a loss of precision. If a loss of precision occurs, a warning is issued. Please note that this also overrides the use of H5TOR_CONV_UNIT64_NA. As loss of precision is already accepted, UINT64-values that are larger than LLONG_MAX will be represented as their next possible floating point value.

H5TOR_CONV_UINT64_NA

When converting an unsigned 64bit integer, any values that don't fit into a signed 64bit integer are set to NA. If this flag is not set, then the values will be truncated to LLONG_MAX, the largest 64bit signed integer.

H5TOR_CONV_DEFAULT

Is both H5TOR_CONV_INT64_INT and H5TOR_CONV_UNIT64_FLOAT

Author

Holger Hoefling

Examples

h5const$overview
#>                              Category
#> 1                             H5F_ACC
#> 2                             H5F_ACC
#> 4                             H5F_ACC
#> 5                             H5F_ACC
#> 6                             H5F_ACC
#> 7                             H5F_ACC
#> 8                             H5F_ACC
#> 9                                 H5P
#> 10                                H5S
#> 11                              H5TOR
#> 12                              H5TOR
#> 13                              H5TOR
#> 14                              H5TOR
#> 15                              H5TOR
#> 16                              H5TOR
#> 17                            H5F_OBJ
#> 18                            H5F_OBJ
#> 19                            H5F_OBJ
#> 20                            H5F_OBJ
#> 21                            H5F_OBJ
#> 22                            H5F_OBJ
#> 23                            H5F_OBJ
#> 24                                H5P
#> 25                                H5P
#> 26                                H5P
#> 27                                H5P
#> 28                                H5P
#> 29                                H5P
#> 30                                H5P
#> 31                                H5P
#> 32                                H5P
#> 33                                H5P
#> 34                                H5P
#> 35                                H5P
#> 36                                H5P
#> 37                                H5P
#> 38                                H5P
#> 39                                H5P
#> 40                         H5Z_FILTER
#> 41                         H5Z_FILTER
#> 42                         H5Z_FILTER
#> 43                         H5Z_FILTER
#> 44                         H5Z_FILTER
#> 45                         H5Z_FILTER
#> 46                         H5Z_FILTER
#> 47                           H5Z_FLAG
#> 48                           H5Z_FLAG
#> 49                      H5P_CRT_ORDER
#> 50                      H5P_CRT_ORDER
#> 51                           H5O_COPY
#> 52                           H5O_COPY
#> 53                           H5O_COPY
#> 54                           H5O_COPY
#> 55                           H5O_COPY
#> 56                           H5O_COPY
#> 57                           H5O_INFO
#> 58                           H5O_INFO
#> 59                           H5O_INFO
#> 60                           H5O_INFO
#> 61                           H5O_INFO
#> 62                           H5O_INFO
#> 63                          H5D_CHUNK
#> 1100                       H5_index_t
#> 2100                       H5_index_t
#> 3                          H5_index_t
#> 410                        H5_index_t
#> 510                   H5_iter_order_t
#> 610                   H5_iter_order_t
#> 710                   H5_iter_order_t
#> 810                   H5_iter_order_t
#> 910                   H5_iter_order_t
#> 1010              H5C_cache_decr_mode
#> 1110              H5C_cache_decr_mode
#> 1210              H5C_cache_decr_mode
#> 1310              H5C_cache_decr_mode
#> 1410        H5C_cache_flash_incr_mode
#> 1510        H5C_cache_flash_incr_mode
#> 1610              H5C_cache_incr_mode
#> 1710              H5C_cache_incr_mode
#> 1810                 H5D_alloc_time_t
#> 1910                 H5D_alloc_time_t
#> 2010                 H5D_alloc_time_t
#> 2110                 H5D_alloc_time_t
#> 2210                 H5D_alloc_time_t
#> 2310                H5D_chunk_index_t
#> 2410                H5D_chunk_index_t
#> 2510                H5D_chunk_index_t
#> 2610                H5D_chunk_index_t
#> 2710                H5D_chunk_index_t
#> 2810                H5D_chunk_index_t
#> 2910                H5D_chunk_index_t
#> 304                   H5D_fill_time_t
#> 311                   H5D_fill_time_t
#> 321                   H5D_fill_time_t
#> 331                   H5D_fill_time_t
#> 341                  H5D_fill_value_t
#> 351                  H5D_fill_value_t
#> 361                  H5D_fill_value_t
#> 371                  H5D_fill_value_t
#> 381                      H5D_layout_t
#> 391                      H5D_layout_t
#> 401                      H5D_layout_t
#> 411                      H5D_layout_t
#> 421                      H5D_layout_t
#> 431                      H5D_layout_t
#> 441  H5D_mpio_actual_chunk_opt_mode_t
#> 451  H5D_mpio_actual_chunk_opt_mode_t
#> 461  H5D_mpio_actual_chunk_opt_mode_t
#> 471         H5D_mpio_actual_io_mode_t
#> 481         H5D_mpio_actual_io_mode_t
#> 491         H5D_mpio_actual_io_mode_t
#> 501         H5D_mpio_actual_io_mode_t
#> 511         H5D_mpio_actual_io_mode_t
#> 521    H5D_mpio_no_collective_cause_t
#> 531    H5D_mpio_no_collective_cause_t
#> 541    H5D_mpio_no_collective_cause_t
#> 551    H5D_mpio_no_collective_cause_t
#> 561    H5D_mpio_no_collective_cause_t
#> 571    H5D_mpio_no_collective_cause_t
#> 581    H5D_mpio_no_collective_cause_t
#> 591    H5D_mpio_no_collective_cause_t
#> 601    H5D_mpio_no_collective_cause_t
#> 611    H5D_mpio_no_collective_cause_t
#> 621                H5D_space_status_t
#> 631                H5D_space_status_t
#> 64                 H5D_space_status_t
#> 65                 H5D_space_status_t
#> 66                     H5D_vds_view_t
#> 67                     H5D_vds_view_t
#> 68                     H5D_vds_view_t
#> 69                    H5E_direction_t
#> 70                    H5E_direction_t
#> 71                         H5E_type_t
#> 72                         H5E_type_t
#> 73                      H5ES_status_t
#> 74                      H5ES_status_t
#> 75                      H5ES_status_t
#> 76                      H5ES_status_t
#> 77                 H5F_close_degree_t
#> 78                 H5F_close_degree_t
#> 79                 H5F_close_degree_t
#> 80                 H5F_close_degree_t
#> 81              H5F_file_space_type_t
#> 82              H5F_file_space_type_t
#> 83              H5F_file_space_type_t
#> 84              H5F_file_space_type_t
#> 85              H5F_file_space_type_t
#> 86              H5F_file_space_type_t
#> 87              H5F_fspace_strategy_t
#> 88              H5F_fspace_strategy_t
#> 89              H5F_fspace_strategy_t
#> 90              H5F_fspace_strategy_t
#> 91              H5F_fspace_strategy_t
#> 92                       H5F_libver_t
#> 93                       H5F_libver_t
#> 94                       H5F_libver_t
#> 95                       H5F_libver_t
#> 96                       H5F_libver_t
#> 97                       H5F_libver_t
#> 98                          H5F_mem_t
#> 99                          H5F_mem_t
#> 100                         H5F_mem_t
#> 101                         H5F_mem_t
#> 102                         H5F_mem_t
#> 103                         H5F_mem_t
#> 104                         H5F_mem_t
#> 105                         H5F_mem_t
#> 106                         H5F_mem_t
#> 107                       H5F_scope_t
#> 108                       H5F_scope_t
#> 109              H5FD_file_image_op_t
#> 110              H5FD_file_image_op_t
#> 111              H5FD_file_image_op_t
#> 112              H5FD_file_image_op_t
#> 113              H5FD_file_image_op_t
#> 114              H5FD_file_image_op_t
#> 115              H5FD_file_image_op_t
#> 116              H5FD_file_image_op_t
#> 117                         H5G_obj_t
#> 118                         H5G_obj_t
#> 119                         H5G_obj_t
#> 120                         H5G_obj_t
#> 121                         H5G_obj_t
#> 122                         H5G_obj_t
#> 123                         H5G_obj_t
#> 124                         H5G_obj_t
#> 125                         H5G_obj_t
#> 126                H5G_storage_type_t
#> 127                H5G_storage_type_t
#> 128                H5G_storage_type_t
#> 129                H5G_storage_type_t
#> 130                        H5I_type_t
#> 131                        H5I_type_t
#> 132                        H5I_type_t
#> 133                        H5I_type_t
#> 134                        H5I_type_t
#> 135                        H5I_type_t
#> 136                        H5I_type_t
#> 137                        H5I_type_t
#> 138                        H5I_type_t
#> 139                        H5I_type_t
#> 140                        H5I_type_t
#> 141                        H5I_type_t
#> 142                        H5I_type_t
#> 143                        H5I_type_t
#> 144                        H5I_type_t
#> 145                        H5I_type_t
#> 146                        H5I_type_t
#> 147                        H5I_type_t
#> 148                        H5L_type_t
#> 149                        H5L_type_t
#> 150                        H5L_type_t
#> 151                        H5L_type_t
#> 152                        H5L_type_t
#> 153                       H5LT_lang_t
#> 154                       H5LT_lang_t
#> 155                       H5LT_lang_t
#> 156                       H5LT_lang_t
#> 157                       H5LT_lang_t
#> 158             H5O_mcdt_search_ret_t
#> 159             H5O_mcdt_search_ret_t
#> 160             H5O_mcdt_search_ret_t
#> 161                        H5O_type_t
#> 162                        H5O_type_t
#> 163                        H5O_type_t
#> 164                        H5O_type_t
#> 165                        H5O_type_t
#> 166                        H5O_type_t
#> 167                       H5PL_type_t
#> 168                       H5PL_type_t
#> 169                       H5PL_type_t
#> 170                       H5PL_type_t
#> 171                        H5R_type_t
#> 172                        H5R_type_t
#> 173                        H5R_type_t
#> 174                        H5R_type_t
#> 175                        H5R_type_t
#> 176                        H5R_type_t
#> 177                        H5R_type_t
#> 178                       H5S_class_t
#> 179                       H5S_class_t
#> 180                       H5S_class_t
#> 181                       H5S_class_t
#> 182                      H5S_sel_type
#> 183                      H5S_sel_type
#> 184                      H5S_sel_type
#> 185                      H5S_sel_type
#> 186                      H5S_sel_type
#> 187                      H5S_sel_type
#> 188                     H5S_seloper_t
#> 189                     H5S_seloper_t
#> 190                     H5S_seloper_t
#> 191                     H5S_seloper_t
#> 192                     H5S_seloper_t
#> 193                     H5S_seloper_t
#> 194                     H5S_seloper_t
#> 195                     H5S_seloper_t
#> 196                     H5S_seloper_t
#> 197                     H5S_seloper_t
#> 198                         H5T_bkg_t
#> 199                         H5T_bkg_t
#> 200                         H5T_bkg_t
#> 201                       H5T_class_t
#> 202                       H5T_class_t
#> 203                       H5T_class_t
#> 204                       H5T_class_t
#> 205                       H5T_class_t
#> 206                       H5T_class_t
#> 207                       H5T_class_t
#> 208                       H5T_class_t
#> 209                       H5T_class_t
#> 210                       H5T_class_t
#> 211                       H5T_class_t
#> 212                       H5T_class_t
#> 213                       H5T_class_t
#> 214                         H5T_cmd_t
#> 215                         H5T_cmd_t
#> 216                         H5T_cmd_t
#> 217                 H5T_conv_except_t
#> 218                 H5T_conv_except_t
#> 219                 H5T_conv_except_t
#> 220                 H5T_conv_except_t
#> 221                 H5T_conv_except_t
#> 222                 H5T_conv_except_t
#> 223                 H5T_conv_except_t
#> 224                    H5T_conv_ret_t
#> 225                    H5T_conv_ret_t
#> 226                    H5T_conv_ret_t
#> 227                        H5T_cset_t
#> 228                        H5T_cset_t
#> 229                        H5T_cset_t
#> 230                        H5T_cset_t
#> 231                        H5T_cset_t
#> 232                        H5T_cset_t
#> 233                        H5T_cset_t
#> 234                        H5T_cset_t
#> 235                        H5T_cset_t
#> 236                        H5T_cset_t
#> 237                        H5T_cset_t
#> 238                        H5T_cset_t
#> 239                        H5T_cset_t
#> 240                        H5T_cset_t
#> 241                        H5T_cset_t
#> 242                        H5T_cset_t
#> 243                        H5T_cset_t
#> 244                   H5T_direction_t
#> 245                   H5T_direction_t
#> 246                   H5T_direction_t
#> 247                        H5T_norm_t
#> 248                        H5T_norm_t
#> 249                        H5T_norm_t
#> 250                        H5T_norm_t
#> 251                       H5T_order_t
#> 252                       H5T_order_t
#> 253                       H5T_order_t
#> 254                       H5T_order_t
#> 255                       H5T_order_t
#> 256                       H5T_order_t
#> 257                         H5T_pad_t
#> 258                         H5T_pad_t
#> 259                         H5T_pad_t
#> 260                         H5T_pad_t
#> 261                         H5T_pad_t
#> 262                        H5T_pers_t
#> 263                        H5T_pers_t
#> 264                        H5T_pers_t
#> 265                        H5T_sign_t
#> 266                        H5T_sign_t
#> 267                        H5T_sign_t
#> 268                        H5T_sign_t
#> 269                         H5T_str_t
#> 270                         H5T_str_t
#> 271                         H5T_str_t
#> 272                         H5T_str_t
#> 273                         H5T_str_t
#> 274                         H5T_str_t
#> 275                         H5T_str_t
#> 276                         H5T_str_t
#> 277                         H5T_str_t
#> 278                         H5T_str_t
#> 279                         H5T_str_t
#> 280                         H5T_str_t
#> 281                         H5T_str_t
#> 282                         H5T_str_t
#> 283                         H5T_str_t
#> 284                         H5T_str_t
#> 285                         H5T_str_t
#> 286                    H5VL_map_get_t
#> 287                    H5VL_map_get_t
#> 288                    H5VL_map_get_t
#> 289                    H5VL_map_get_t
#> 290                    H5VL_map_get_t
#> 291               H5VL_map_specific_t
#> 292               H5VL_map_specific_t
#> 293                   H5Z_cb_return_t
#> 294                   H5Z_cb_return_t
#> 295                   H5Z_cb_return_t
#> 296                   H5Z_cb_return_t
#> 297                         H5Z_EDC_t
#> 298                         H5Z_EDC_t
#> 299                         H5Z_EDC_t
#> 300                         H5Z_EDC_t
#> 301               H5Z_SO_scale_type_t
#> 302               H5Z_SO_scale_type_t
#> 303               H5Z_SO_scale_type_t
#>                                                   Name           Constant
#> 1                                        H5F_ACC_TRUNC                  2
#> 2                                         H5F_ACC_EXCL                  4
#> 4                                       H5F_ACC_RDONLY                  0
#> 5                                         H5F_ACC_RDWR                  1
#> 6                                      H5F_ACC_DEFAULT              65535
#> 7                                   H5F_ACC_SWMR_WRITE                 32
#> 8                                    H5F_ACC_SWMR_READ                 64
#> 9                                          H5P_DEFAULT                  0
#> 10                                             H5S_ALL                  0
#> 11                                     H5TOR_CONV_NONE                  0
#> 12                         H5TOR_CONV_INT64_INT_NOLOSS                  1
#> 13                       H5TOR_CONV_INT64_FLOAT_NOLOSS                  2
#> 14                        H5TOR_CONV_INT64_FLOAT_FORCE                  4
#> 15                                H5TOR_CONV_UINT64_NA                  8
#> 16                             H5TOR_CONV_INT64_NOLOSS                  3
#> 17                                        H5F_OBJ_FILE                  1
#> 18                                     H5F_OBJ_DATASET                  2
#> 19                                       H5F_OBJ_GROUP                  4
#> 20                                    H5F_OBJ_DATATYPE                  8
#> 21                                        H5F_OBJ_ATTR                 16
#> 22                                         H5F_OBJ_ALL                 31
#> 23                                       H5F_OBJ_LOCAL                 32
#> 24                                H5P_ATTRIBUTE_CREATE 720575940379279377
#> 25                                  H5P_DATASET_ACCESS 720575940379279370
#> 26                                  H5P_DATASET_CREATE 720575940379279369
#> 27                                    H5P_DATASET_XFER 720575940379279371
#> 28                                 H5P_DATATYPE_ACCESS 720575940379279374
#> 29                                 H5P_DATATYPE_CREATE 720575940379279373
#> 30                                     H5P_FILE_ACCESS 720575940379279368
#> 31                                     H5P_FILE_CREATE 720575940379279367
#> 32                                      H5P_FILE_MOUNT 720575940379279372
#> 33                                    H5P_GROUP_ACCESS 720575940379279366
#> 34                                    H5P_GROUP_CREATE 720575940379279364
#> 35                                     H5P_LINK_ACCESS 720575940379279363
#> 36                                     H5P_LINK_CREATE 720575940379279379
#> 37                                     H5P_OBJECT_COPY 720575940379279365
#> 38                                   H5P_OBJECT_CREATE 720575940379279361
#> 39                                   H5P_STRING_CREATE 720575940379279362
#> 40                                      H5Z_FILTER_ALL                  0
#> 41                                  H5Z_FILTER_DEFLATE                  1
#> 42                                  H5Z_FILTER_SHUFFLE                  2
#> 43                               H5Z_FILTER_FLETCHER32                  3
#> 44                                     H5Z_FILTER_SZIP                  4
#> 45                                     H5Z_FILTER_NBIT                  5
#> 46                              H5Z_FILTER_SCALEOFFSET                  6
#> 47                                   H5Z_FLAG_OPTIONAL                  1
#> 48                                  H5Z_FLAG_MANDATORY                  0
#> 49                               H5P_CRT_ORDER_TRACKED                  1
#> 50                               H5P_CRT_ORDER_INDEXED                  2
#> 51                     H5O_COPY_SHALLOW_HIERARCHY_FLAG                  1
#> 52                      H5O_COPY_EXPAND_SOFT_LINK_FLAG                  2
#> 53                       H5O_COPY_EXPAND_EXT_LINK_FLAG                  4
#> 54                      H5O_COPY_EXPAND_REFERENCE_FLAG                  8
#> 55                          H5O_COPY_WITHOUT_ATTR_FLAG                 16
#> 56                 H5O_COPY_MERGE_COMMITTED_DTYPE_FLAG                 64
#> 57                                      H5O_INFO_BASIC                  1
#> 58                                       H5O_INFO_TIME                  2
#> 59                                  H5O_INFO_NUM_ATTRS                  4
#> 60                                        H5O_INFO_HDR                  8
#> 61                                  H5O_INFO_META_SIZE                 16
#> 62                                        H5O_INFO_ALL                 31
#> 63                          H5D_CHUNK_CACHE_W0_DEFAULT                 -1
#> 1100                                  H5_INDEX_UNKNOWN                 -1
#> 2100                                     H5_INDEX_NAME                  0
#> 3                                   H5_INDEX_CRT_ORDER                  1
#> 410                                         H5_INDEX_N                  2
#> 510                                    H5_ITER_UNKNOWN                 -1
#> 610                                        H5_ITER_INC                  0
#> 710                                        H5_ITER_DEC                  1
#> 810                                     H5_ITER_NATIVE                  2
#> 910                                          H5_ITER_N                  3
#> 1010                                     H5C_decr__off                  0
#> 1110                               H5C_decr__threshold                  1
#> 1210                                 H5C_decr__age_out                  2
#> 1310                  H5C_decr__age_out_with_threshold                  3
#> 1410                               H5C_flash_incr__off                  0
#> 1510                         H5C_flash_incr__add_space                  1
#> 1610                                     H5C_incr__off                  0
#> 1710                               H5C_incr__threshold                  1
#> 1810                              H5D_ALLOC_TIME_ERROR                 -1
#> 1910                            H5D_ALLOC_TIME_DEFAULT                  0
#> 2010                              H5D_ALLOC_TIME_EARLY                  1
#> 2110                               H5D_ALLOC_TIME_LATE                  2
#> 2210                               H5D_ALLOC_TIME_INCR                  3
#> 2310                               H5D_CHUNK_IDX_BTREE                  0
#> 2410                              H5D_CHUNK_IDX_SINGLE                  1
#> 2510                                H5D_CHUNK_IDX_NONE                  2
#> 2610                              H5D_CHUNK_IDX_FARRAY                  3
#> 2710                              H5D_CHUNK_IDX_EARRAY                  4
#> 2810                                 H5D_CHUNK_IDX_BT2                  5
#> 2910                              H5D_CHUNK_IDX_NTYPES                  6
#> 304                                H5D_FILL_TIME_ERROR                 -1
#> 311                                H5D_FILL_TIME_ALLOC                  0
#> 321                                H5D_FILL_TIME_NEVER                  1
#> 331                                H5D_FILL_TIME_IFSET                  2
#> 341                               H5D_FILL_VALUE_ERROR                 -1
#> 351                           H5D_FILL_VALUE_UNDEFINED                  0
#> 361                             H5D_FILL_VALUE_DEFAULT                  1
#> 371                        H5D_FILL_VALUE_USER_DEFINED                  2
#> 381                                   H5D_LAYOUT_ERROR                 -1
#> 391                                        H5D_COMPACT                  0
#> 401                                     H5D_CONTIGUOUS                  1
#> 411                                        H5D_CHUNKED                  2
#> 421                                        H5D_VIRTUAL                  3
#> 431                                       H5D_NLAYOUTS                  4
#> 441                     H5D_MPIO_NO_CHUNK_OPTIMIZATION                  0
#> 451                                H5D_MPIO_LINK_CHUNK                  1
#> 461                               H5D_MPIO_MULTI_CHUNK                  2
#> 471                             H5D_MPIO_NO_COLLECTIVE                  0
#> 481                         H5D_MPIO_CHUNK_INDEPENDENT                  1
#> 491                          H5D_MPIO_CHUNK_COLLECTIVE                  2
#> 501                               H5D_MPIO_CHUNK_MIXED                  3
#> 511                     H5D_MPIO_CONTIGUOUS_COLLECTIVE                  4
#> 521                                H5D_MPIO_COLLECTIVE                  0
#> 531                           H5D_MPIO_SET_INDEPENDENT                  1
#> 541                       H5D_MPIO_DATATYPE_CONVERSION                  2
#> 551                           H5D_MPIO_DATA_TRANSFORMS                  4
#> 561            H5D_MPIO_MPI_OPT_TYPES_ENV_VAR_DISABLED                  8
#> 571           H5D_MPIO_NOT_SIMPLE_OR_SCALAR_DATASPACES                 16
#> 581         H5D_MPIO_NOT_CONTIGUOUS_OR_CHUNKED_DATASET                 32
#> 591         H5D_MPIO_PARALLEL_FILTERED_WRITES_DISABLED                 64
#> 601  H5D_MPIO_ERROR_WHILE_CHECKING_COLLECTIVE_POSSIBLE                128
#> 611                   H5D_MPIO_NO_COLLECTIVE_MAX_CAUSE                256
#> 621                             H5D_SPACE_STATUS_ERROR                 -1
#> 631                     H5D_SPACE_STATUS_NOT_ALLOCATED                  0
#> 64                     H5D_SPACE_STATUS_PART_ALLOCATED                  1
#> 65                          H5D_SPACE_STATUS_ALLOCATED                  2
#> 66                                       H5D_VDS_ERROR                 -1
#> 67                               H5D_VDS_FIRST_MISSING                  0
#> 68                              H5D_VDS_LAST_AVAILABLE                  1
#> 69                                     H5E_WALK_UPWARD                  0
#> 70                                   H5E_WALK_DOWNWARD                  1
#> 71                                           H5E_MAJOR                  0
#> 72                                           H5E_MINOR                  1
#> 73                             H5ES_STATUS_IN_PROGRESS                  0
#> 74                                 H5ES_STATUS_SUCCEED                  1
#> 75                                    H5ES_STATUS_FAIL                  2
#> 76                                H5ES_STATUS_CANCELED                  3
#> 77                                   H5F_CLOSE_DEFAULT                  0
#> 78                                      H5F_CLOSE_WEAK                  1
#> 79                                      H5F_CLOSE_SEMI                  2
#> 80                                    H5F_CLOSE_STRONG                  3
#> 81                              H5F_FILE_SPACE_DEFAULT                  0
#> 82                          H5F_FILE_SPACE_ALL_PERSIST                  1
#> 83                                  H5F_FILE_SPACE_ALL                  2
#> 84                             H5F_FILE_SPACE_AGGR_VFD                  3
#> 85                                  H5F_FILE_SPACE_VFD                  4
#> 86                               H5F_FILE_SPACE_NTYPES                  5
#> 87                        H5F_FSPACE_STRATEGY_FSM_AGGR                  0
#> 88                            H5F_FSPACE_STRATEGY_PAGE                  1
#> 89                            H5F_FSPACE_STRATEGY_AGGR                  2
#> 90                            H5F_FSPACE_STRATEGY_NONE                  3
#> 91                          H5F_FSPACE_STRATEGY_NTYPES                  4
#> 92                                    H5F_LIBVER_ERROR                 -1
#> 93                                 H5F_LIBVER_EARLIEST                  0
#> 94                                      H5F_LIBVER_V18                  1
#> 95                                     H5F_LIBVER_V110                  2
#> 96                                     H5F_LIBVER_V112                  3
#> 97                                  H5F_LIBVER_NBOUNDS                  4
#> 98                                     H5FD_MEM_NOLIST                 -1
#> 99                                    H5FD_MEM_DEFAULT                  0
#> 100                                     H5FD_MEM_SUPER                  1
#> 101                                     H5FD_MEM_BTREE                  2
#> 102                                      H5FD_MEM_DRAW                  3
#> 103                                     H5FD_MEM_GHEAP                  4
#> 104                                     H5FD_MEM_LHEAP                  5
#> 105                                      H5FD_MEM_OHDR                  6
#> 106                                    H5FD_MEM_NTYPES                  7
#> 107                                    H5F_SCOPE_LOCAL                  0
#> 108                                   H5F_SCOPE_GLOBAL                  1
#> 109                           H5FD_FILE_IMAGE_OP_NO_OP                  0
#> 110               H5FD_FILE_IMAGE_OP_PROPERTY_LIST_SET                  1
#> 111              H5FD_FILE_IMAGE_OP_PROPERTY_LIST_COPY                  2
#> 112               H5FD_FILE_IMAGE_OP_PROPERTY_LIST_GET                  3
#> 113             H5FD_FILE_IMAGE_OP_PROPERTY_LIST_CLOSE                  4
#> 114                       H5FD_FILE_IMAGE_OP_FILE_OPEN                  5
#> 115                     H5FD_FILE_IMAGE_OP_FILE_RESIZE                  6
#> 116                      H5FD_FILE_IMAGE_OP_FILE_CLOSE                  7
#> 117                                        H5G_UNKNOWN                 -1
#> 118                                          H5G_GROUP                  0
#> 119                                        H5G_DATASET                  1
#> 120                                           H5G_TYPE                  2
#> 121                                           H5G_LINK                  3
#> 122                                         H5G_UDLINK                  4
#> 123                                     H5G_RESERVED_5                  5
#> 124                                     H5G_RESERVED_6                  6
#> 125                                     H5G_RESERVED_7                  7
#> 126                           H5G_STORAGE_TYPE_UNKNOWN                 -1
#> 127                      H5G_STORAGE_TYPE_SYMBOL_TABLE                  0
#> 128                           H5G_STORAGE_TYPE_COMPACT                  1
#> 129                             H5G_STORAGE_TYPE_DENSE                  2
#> 130                                         H5I_UNINIT                 -2
#> 131                                          H5I_BADID                 -1
#> 132                                           H5I_FILE                  1
#> 133                                          H5I_GROUP                  2
#> 134                                       H5I_DATATYPE                  3
#> 135                                      H5I_DATASPACE                  4
#> 136                                        H5I_DATASET                  5
#> 137                                            H5I_MAP                  6
#> 138                                           H5I_ATTR                  7
#> 139                                            H5I_VFL                  8
#> 140                                            H5I_VOL                  9
#> 141                                    H5I_GENPROP_CLS                 10
#> 142                                    H5I_GENPROP_LST                 11
#> 143                                    H5I_ERROR_CLASS                 12
#> 144                                      H5I_ERROR_MSG                 13
#> 145                                    H5I_ERROR_STACK                 14
#> 146                                 H5I_SPACE_SEL_ITER                 15
#> 147                                         H5I_NTYPES                 16
#> 148                                     H5L_TYPE_ERROR                 -1
#> 149                                      H5L_TYPE_HARD                  0
#> 150                                      H5L_TYPE_SOFT                  1
#> 151                                  H5L_TYPE_EXTERNAL                 64
#> 152                                       H5L_TYPE_MAX                255
#> 153                                      H5LT_LANG_ERR                 -1
#> 154                                           H5LT_DDL                  0
#> 155                                             H5LT_C                  1
#> 156                                       H5LT_FORTRAN                  2
#> 157                                       H5LT_NO_LANG                  3
#> 158                              H5O_MCDT_SEARCH_ERROR                 -1
#> 159                               H5O_MCDT_SEARCH_CONT                  0
#> 160                               H5O_MCDT_SEARCH_STOP                  1
#> 161                                   H5O_TYPE_UNKNOWN                 -1
#> 162                                     H5O_TYPE_GROUP                  0
#> 163                                   H5O_TYPE_DATASET                  1
#> 164                            H5O_TYPE_NAMED_DATATYPE                  2
#> 165                                       H5O_TYPE_MAP                  3
#> 166                                    H5O_TYPE_NTYPES                  4
#> 167                                    H5PL_TYPE_ERROR                 -1
#> 168                                   H5PL_TYPE_FILTER                  0
#> 169                                      H5PL_TYPE_VOL                  1
#> 170                                     H5PL_TYPE_NONE                  2
#> 171                                        H5R_BADTYPE                 -1
#> 172                                        H5R_OBJECT1                  0
#> 173                                H5R_DATASET_REGION1                  1
#> 174                                        H5R_OBJECT2                  2
#> 175                                H5R_DATASET_REGION2                  3
#> 176                                           H5R_ATTR                  4
#> 177                                        H5R_MAXTYPE                  5
#> 178                                       H5S_NO_CLASS                 -1
#> 179                                         H5S_SCALAR                  0
#> 180                                         H5S_SIMPLE                  1
#> 181                                           H5S_NULL                  2
#> 182                                      H5S_SEL_ERROR                 -1
#> 183                                       H5S_SEL_NONE                  0
#> 184                                     H5S_SEL_POINTS                  1
#> 185                                 H5S_SEL_HYPERSLABS                  2
#> 186                                        H5S_SEL_ALL                  3
#> 187                                          H5S_SEL_N                  4
#> 188                                    H5S_SELECT_NOOP                 -1
#> 189                                     H5S_SELECT_SET                  0
#> 190                                      H5S_SELECT_OR                  1
#> 191                                     H5S_SELECT_AND                  2
#> 192                                     H5S_SELECT_XOR                  3
#> 193                                    H5S_SELECT_NOTB                  4
#> 194                                    H5S_SELECT_NOTA                  5
#> 195                                  H5S_SELECT_APPEND                  6
#> 196                                 H5S_SELECT_PREPEND                  7
#> 197                                 H5S_SELECT_INVALID                  8
#> 198                                         H5T_BKG_NO                  0
#> 199                                       H5T_BKG_TEMP                  1
#> 200                                        H5T_BKG_YES                  2
#> 201                                       H5T_NO_CLASS                 -1
#> 202                                        H5T_INTEGER                  0
#> 203                                          H5T_FLOAT                  1
#> 204                                           H5T_TIME                  2
#> 205                                         H5T_STRING                  3
#> 206                                       H5T_BITFIELD                  4
#> 207                                         H5T_OPAQUE                  5
#> 208                                       H5T_COMPOUND                  6
#> 209                                      H5T_REFERENCE                  7
#> 210                                           H5T_ENUM                  8
#> 211                                           H5T_VLEN                  9
#> 212                                          H5T_ARRAY                 10
#> 213                                       H5T_NCLASSES                 11
#> 214                                      H5T_CONV_INIT                  0
#> 215                                      H5T_CONV_CONV                  1
#> 216                                      H5T_CONV_FREE                  2
#> 217                           H5T_CONV_EXCEPT_RANGE_HI                  0
#> 218                          H5T_CONV_EXCEPT_RANGE_LOW                  1
#> 219                          H5T_CONV_EXCEPT_PRECISION                  2
#> 220                           H5T_CONV_EXCEPT_TRUNCATE                  3
#> 221                               H5T_CONV_EXCEPT_PINF                  4
#> 222                               H5T_CONV_EXCEPT_NINF                  5
#> 223                                H5T_CONV_EXCEPT_NAN                  6
#> 224                                     H5T_CONV_ABORT                 -1
#> 225                                 H5T_CONV_UNHANDLED                  0
#> 226                                   H5T_CONV_HANDLED                  1
#> 227                                     H5T_CSET_ERROR                 -1
#> 228                                     H5T_CSET_ASCII                  0
#> 229                                      H5T_CSET_UTF8                  1
#> 230                                H5T_CSET_RESERVED_2                  2
#> 231                                H5T_CSET_RESERVED_3                  3
#> 232                                H5T_CSET_RESERVED_4                  4
#> 233                                H5T_CSET_RESERVED_5                  5
#> 234                                H5T_CSET_RESERVED_6                  6
#> 235                                H5T_CSET_RESERVED_7                  7
#> 236                                H5T_CSET_RESERVED_8                  8
#> 237                                H5T_CSET_RESERVED_9                  9
#> 238                               H5T_CSET_RESERVED_10                 10
#> 239                               H5T_CSET_RESERVED_11                 11
#> 240                               H5T_CSET_RESERVED_12                 12
#> 241                               H5T_CSET_RESERVED_13                 13
#> 242                               H5T_CSET_RESERVED_14                 14
#> 243                               H5T_CSET_RESERVED_15                 15
#> 244                                    H5T_DIR_DEFAULT                  0
#> 245                                     H5T_DIR_ASCEND                  1
#> 246                                    H5T_DIR_DESCEND                  2
#> 247                                     H5T_NORM_ERROR                 -1
#> 248                                   H5T_NORM_IMPLIED                  0
#> 249                                    H5T_NORM_MSBSET                  1
#> 250                                      H5T_NORM_NONE                  2
#> 251                                    H5T_ORDER_ERROR                 -1
#> 252                                       H5T_ORDER_LE                  0
#> 253                                       H5T_ORDER_BE                  1
#> 254                                      H5T_ORDER_VAX                  2
#> 255                                    H5T_ORDER_MIXED                  3
#> 256                                     H5T_ORDER_NONE                  4
#> 257                                      H5T_PAD_ERROR                 -1
#> 258                                       H5T_PAD_ZERO                  0
#> 259                                        H5T_PAD_ONE                  1
#> 260                                 H5T_PAD_BACKGROUND                  2
#> 261                                           H5T_NPAD                  3
#> 262                                  H5T_PERS_DONTCARE                 -1
#> 263                                      H5T_PERS_HARD                  0
#> 264                                      H5T_PERS_SOFT                  1
#> 265                                      H5T_SGN_ERROR                 -1
#> 266                                       H5T_SGN_NONE                  0
#> 267                                          H5T_SGN_2                  1
#> 268                                           H5T_NSGN                  2
#> 269                                      H5T_STR_ERROR                 -1
#> 270                                   H5T_STR_NULLTERM                  0
#> 271                                    H5T_STR_NULLPAD                  1
#> 272                                   H5T_STR_SPACEPAD                  2
#> 273                                 H5T_STR_RESERVED_3                  3
#> 274                                 H5T_STR_RESERVED_4                  4
#> 275                                 H5T_STR_RESERVED_5                  5
#> 276                                 H5T_STR_RESERVED_6                  6
#> 277                                 H5T_STR_RESERVED_7                  7
#> 278                                 H5T_STR_RESERVED_8                  8
#> 279                                 H5T_STR_RESERVED_9                  9
#> 280                                H5T_STR_RESERVED_10                 10
#> 281                                H5T_STR_RESERVED_11                 11
#> 282                                H5T_STR_RESERVED_12                 12
#> 283                                H5T_STR_RESERVED_13                 13
#> 284                                H5T_STR_RESERVED_14                 14
#> 285                                H5T_STR_RESERVED_15                 15
#> 286                                  H5VL_MAP_GET_MAPL                  0
#> 287                                  H5VL_MAP_GET_MCPL                  1
#> 288                              H5VL_MAP_GET_KEY_TYPE                  2
#> 289                              H5VL_MAP_GET_VAL_TYPE                  3
#> 290                                 H5VL_MAP_GET_COUNT                  4
#> 291                                      H5VL_MAP_ITER                  0
#> 292                                    H5VL_MAP_DELETE                  1
#> 293                                       H5Z_CB_ERROR                 -1
#> 294                                        H5Z_CB_FAIL                  0
#> 295                                        H5Z_CB_CONT                  1
#> 296                                          H5Z_CB_NO                  2
#> 297                                      H5Z_ERROR_EDC                 -1
#> 298                                    H5Z_DISABLE_EDC                  0
#> 299                                     H5Z_ENABLE_EDC                  1
#> 300                                         H5Z_NO_EDC                  2
#> 301                                H5Z_SO_FLOAT_DSCALE                  0
#> 302                                H5Z_SO_FLOAT_ESCALE                  1
#> 303                                         H5Z_SO_INT                  2
h5const$H5F_ACC_RDWR
#> [1] H5F_ACC_RDWR
#> 7 Levels: H5F_ACC_RDONLY H5F_ACC_RDWR H5F_ACC_TRUNC ... H5F_ACC_DEFAULT
#> 7 Values: 0 1 2 ... 65535
h5const$H5F_ACC_DEFAULT
#> [1] H5F_ACC_DEFAULT
#> 7 Levels: H5F_ACC_RDONLY H5F_ACC_RDWR H5F_ACC_TRUNC ... H5F_ACC_DEFAULT
#> 7 Values: 0 1 2 ... 65535
# Combining flags
bitwOr(h5const$H5TOR_CONV_UINT64_NA, h5const$H5TOR_CONV_INT64_INT_NOLOSS)
#> [1] 9